% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.r
\name{add_metadata}
\alias{add_metadata}
\title{Add meta data to extracted data}
\usage{
add_metadata(dat, cols = c("sample_size", "ncase", "ncontrol", "unit", "sd"))
}
\arguments{
\item{dat}{Either exposure data, outcome data or harmonised data}

\item{cols}{Which metadata fields to add. Default = c("sample_size", "ncase", "ncontrol", "unit", "sd")}
}
\value{
Data frame
}
\description{
Previously the meta data was returned alongside association information. This is mostly unnecessary as it is needlessly repeating information. This is a convenience function that reinstates that information.
Can be applied to either exposure data, outcome data, or harmonised data
}
