% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rsq.r
\name{add_rsq}
\alias{add_rsq}
\title{Estimate r-square of each association}
\usage{
add_rsq(dat)
}
\arguments{
\item{dat}{exposure_dat, outcome_dat or harmonised_data}
}
\value{
data frame
}
\description{
Can be applied to exposure_dat, outcome_dat or harmonised_data. Note that it will be beneficial in some circumstances to add the meta data to the data object using \code{add_metadata()} before running this function. Also adds effective sample size for case control data
}
