% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mysql.R
\name{available_outcomes}
\alias{available_outcomes}
\title{Get list of studies with available GWAS summary statistics}
\usage{
available_outcomes(user = "mruser", password = "TMG_F1WnTL",
  dbname = "mrbase", host = "epi-franklin.epi.bris.ac.uk")
}
\arguments{
\item{user="mruser"}{User name}

\item{password="TMG_F1WnTL"}{Password}

\item{dbname="mrbase"}{Database}

\item{host="epi-franklin.epi.bris.ac.uk"}{Host}
}
\value{
Dataframe of study details
}
\description{
Get list of studies with available GWAS summary statistics
}

