% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{available_outcomes_mysql}
\alias{available_outcomes_mysql}
\title{Get list of studies with available GWAS summary statistics}
\usage{
available_outcomes_mysql(user, password, dbname, host, port)
}
\arguments{
\item{user}{User name}

\item{password}{Password}

\item{dbname}{Database}

\item{host}{Host}
}
\value{
Dataframe of study details
}
\description{
Get list of studies with available GWAS summary statistics
}
