% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{check_bcftools}
\alias{check_bcftools}
\title{Check if bcftools is available on the path}
\usage{
check_bcftools()
}
\value{
Status code from running `bcftools --help`. Expect 0 if bcftools available
}
\description{
Check if bcftools is available on the path
}
