% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_mrbase_access}
\alias{check_mrbase_access}
\title{Check MR Base access level}
\usage{
check_mrbase_access()
}
\value{
access level string
}
\description{
In order to be granted access to a particular dataset that is not generally available please contact the developers.
}
