% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\name{check_password}
\alias{check_password}
\title{Check password for MR Base API}
\usage{
check_password(password)
}
\arguments{
\item{password}{String}
}
\value{
Boolean whether password is accepted or not.
}
\description{
Check password for MR Base API
}
