% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{cleanup_outcome_data}
\alias{cleanup_outcome_data}
\title{Avoid issues in MR by finding impossible vals and setting to NA}
\usage{
cleanup_outcome_data(d)
}
\arguments{
\item{d}{Data frame}
}
\value{
Cleaned data frame
}
\description{
Avoid issues in MR by finding impossible vals and setting to NA
}
\keyword{internal}
