% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{clump_data}
\alias{clump_data}
\title{Perform LD clumping on SNP data}
\usage{
clump_data(dat, where = "remote", refdat = NULL, plink_bin = NULL,
  clump_kb = 10000, clump_r2 = 0.1, clump_p1 = 1, clump_p2 = 1,
  tempdir = getwd())
}
\arguments{
\item{dat}{Output from \code{read_exposure_data}. Must have a SNP name column (SNP), SNP chromosome column (chr_name), SNP position column (chrom_start) and p-value column (pval.exposure)}

\item{where}{Where to perform clumping - if using MR-Base servers then specify "remote" (default), if providing your own reference data and plink binary then specify "local"}

\item{refdat}{Path to binary plink files to use for LD referencing. If NULL (default) and where="local" then looks for data provided by the Eur1000Genomes R package.}

\item{plink_bin}{Path to plink binary. If NULL (default) and where="local" then looks for executable in Eur1000Genomes R package.}

\item{tempdir}{= "." Where to print results}

\item{clump_kb=10000}{Clumping window}

\item{clump_r2=0.1}{Clumping r2 cutoff}

\item{clump_p1=1}{Clumping sig level for index SNPs}

\item{clump_p2=1}{Clumping sig level for secondary SNPs}
}
\value{
Data frame
}
\description{
Uses PLINK clumping method, where SNPs in LD within a particular window will be pruned. The SNP with the lowest p-value is retained.
}

