% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr_results2.R
\name{combine_all_mrresults}
\alias{combine_all_mrresults}
\title{Combine all mr results}
\usage{
combine_all_mrresults(
  res,
  het,
  plt,
  sin,
  ao_slc = TRUE,
  Exp = FALSE,
  split.exposure = FALSE,
  split.outcome = FALSE
)
}
\arguments{
\item{res}{Results from \code{\link[=mr]{mr()}}.}

\item{het}{Results from \code{\link[=mr_heterogeneity]{mr_heterogeneity()}}.}

\item{plt}{Results from \code{\link[=mr_pleiotropy_test]{mr_pleiotropy_test()}}.}

\item{sin}{Results from \code{\link[=mr_singlesnp]{mr_singlesnp()}}.}

\item{ao_slc}{Logical; if set to \code{TRUE} then outcome study level characteristics are retrieved from \code{\link[=available_outcomes]{available_outcomes()}}. Default is \code{TRUE}.}

\item{Exp}{Logical; if set to \code{TRUE} results are exponentiated. Useful if user wants log odds ratios expressed as odds ratios. Default is \code{FALSE}.}

\item{split.exposure}{Logical; if set to \code{TRUE} the exposure column is split into separate columns for the exposure name and exposure ID. Default is \code{FALSE}.}

\item{split.outcome}{Logical; if set to \code{TRUE} the outcome column is split into separate columns for the outcome name and outcome ID. Default is \code{FALSE}.}
}
\value{
data frame
}
\description{
This function combines results of \code{\link[=mr]{mr()}}, \code{\link[=mr_heterogeneity]{mr_heterogeneity()}}, \code{\link[=mr_pleiotropy_test]{mr_pleiotropy_test()}} and \code{\link[=mr_singlesnp]{mr_singlesnp()}} into a single data frame.
It also merges the results with outcome study level characteristics in \code{\link[=available_outcomes]{available_outcomes()}}.
If desired it also exponentiates results (e.g. if the user wants log odds ratio converted into odds ratios with 95 percent confidence intervals).
The exposure and outcome columns from the output from \code{\link[=mr]{mr()}} contain both the trait names and trait ids.
The \code{combine_all_mrresults()} function splits these into separate columns by default.
}
