% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{combine_data}
\alias{combine_data}
\title{Combine data}
\usage{
combine_data(x)
}
\arguments{
\item{x}{List of data frames returned from \code{\link{format_data}}.}
}
\value{
data frame
}
\description{
Taking exposure or outcome data (returned from \code{\link{format_data}})
combine multiple datasets together so they can be analysed in one
batch. Removes duplicate SNPs, preferentially keeping those usable
in MR analysis.
}
