% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eve.R
\name{combine_ml}
\alias{combine_ml}
\title{Combine two ml objects to create a harmonised dataset}
\usage{
combine_ml(ml1, ml2)
}
\arguments{
\item{ml1}{Output from read_ml}

\item{ml2}{Output from read_ml}
}
\value{
data frame
}
\description{
Combine two ml objects to create a harmonised dataset
}
