% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rsq.r
\name{contingency}
\alias{contingency}
\title{Obtain 2x2 contingency table from marginal parameters and odds ratio}
\usage{
contingency(af, prop, odds_ratio, eps = 1e-15)
}
\arguments{
\item{af}{Allele frequency of effect allele.}

\item{prop}{Proportion of cases.}

\item{odds_ratio}{Odds ratio.}

\item{eps}{tolerance, default is \code{1e-15}.}
}
\value{
2x2 contingency table as matrix
}
\description{
Columns are the case and control frequencies.
Rows are the frequencies for allele 1 and allele 2.
}
