% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R, R/read_data.R
\name{convert_outcome_to_exposure}
\alias{convert_outcome_to_exposure}
\title{Convert outcome format to exposure format}
\usage{
convert_outcome_to_exposure(outcome_dat)

convert_outcome_to_exposure(outcome_dat)
}
\arguments{
\item{outcome_dat}{Output from \code{extract_outcome_data}}
}
\value{
Data frame

data frame
}
\description{
Helper function to convert results from extract_outcome_data to exposure_dat format
}
