% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot2.R
\name{create_label}
\alias{create_label}
\title{Create fixed width label}
\usage{
create_label(n1, nom)
}
\arguments{
\item{n1}{number}

\item{nom}{name}
}
\value{
text
}
\description{
Create fixed width label
}
\keyword{internal}
