% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{dat_to_MRInput}
\alias{dat_to_MRInput}
\title{Convert TwoSampleMR format to MendelianRandomization format}
\usage{
dat_to_MRInput(dat)
}
\arguments{
\item{dat}{Output from the \code{harmonise_data} function}
}
\value{
List of MRInput objects for each exposure/outcome combination
}
\description{
The MendelianRandomization package offers MR methods that can 
be used with the same data used in the TwoSampleMR package. This
function converts from the TwoSampleMR format to the MRInput class.
}
