% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_formats.R
\name{dat_to_RadialMR}
\alias{dat_to_RadialMR}
\title{Convert dat to RadialMR format}
\usage{
dat_to_RadialMR(dat)
}
\arguments{
\item{dat}{Output from \code{harmonise_data}}
}
\value{
List of RadialMR format datasets
}
\description{
Creates a list of RadialMR format datasets for each exposure - outcome pair
}
