% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{default_parameters}
\alias{default_parameters}
\title{List of parameters for use with MR functions}
\usage{
default_parameters()
}
\description{
The default is \code{list(test_dist = "z", nboot = 1000, Cov = 0, penk = 20, phi = 1, alpha = 0.05, Qthresh = 0.05, over.dispersion = TRUE, loss.function = "huber")}.
}
