% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steiger.R
\name{directionality_test}
\alias{directionality_test}
\title{Perform MR Steiger test of directionality}
\usage{
directionality_test(dat)
}
\arguments{
\item{dat}{Harmonised exposure and outcome data. Output from \code{\link{harmonise_data}}.}
}
\value{
List
}
\description{
A statistical test for whether the assumption that exposure causes outcome is valid.
}
