% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rsq.r
\name{effective_n}
\alias{effective_n}
\title{Estimate the effective sample size in a case control study}
\usage{
effective_n(ncase, ncontrol)
}
\arguments{
\item{ncase}{Vector of number of cases}

\item{ncontrol}{Vector of number of controls}
}
\value{
Vector of effective sample size
}
\description{
Taken from https://www.nature.com/articles/nprot.2014.071
}
