% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mr.R
\name{eggers_regression_bootstrap}
\alias{eggers_regression_bootstrap}
\title{Run bootstrap to generate standard errors for MR}
\usage{
eggers_regression_bootstrap(b_exp, b_out, se_exp, se_out, nboot = 1000,
  alpha = 0.05)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}

\item{nboot}{Number of bootstraps. Default 1000}

\item{alpha}{Quantile to use for confidence intervals. Default 0.05}
}
\value{
A list with the following elements:
        boots: data frame of bootstrap results
        res: data frame of summary results from each bootstrap
        data: data frame of input summary stats
}
\description{
Run bootstrap to generate standard errors for MR
}

