% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment_method_list}
\alias{enrichment_method_list}
\title{Get list of available p-value enrichment methods}
\usage{
enrichment_method_list()
}
\value{
Data frame
}
\description{
Get list of available p-value enrichment methods
}
