% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{estimate_trait_sd}
\alias{estimate_trait_sd}
\title{Estimate trait SD by obtaining beta estimates from z-scores and finding the ratio with original beta values}
\usage{
estimate_trait_sd(b, se, n, p)
}
\arguments{
\item{b}{vector of effect sizes}

\item{se}{vector of standard errors}

\item{n}{vector of sample sizes}

\item{p}{vector of allele frequencies}
}
\value{
vector of sd estimates for each association
}
\description{
Assumes that sample size and allele frequency is correct for each SNP, and that allele frequency gives a reasonable estimate of the variance of the SNP
}
