% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{extract_instruments}
\alias{extract_instruments}
\title{Find instruments for use in MR from the MR Base database}
\usage{
extract_instruments(outcomes, p1 = 5e-08, clump = TRUE, p2 = 5e-08,
  r2 = 0.001, kb = 10000, access_token = get_mrbase_access_token(),
  force_server = FALSE, force_server_if_empty = TRUE)
}
\arguments{
\item{outcomes}{Array of outcome IDs (see \code{available_outcomes})}

\item{p1}{= 5e-8 Significance threshold}

\item{clump}{= TRUE Whether to clump results}

\item{p2}{= 5e-8 Secondary clumping threshold}

\item{r2}{= 0.001 Clumping r2 cut off}

\item{kb}{= 10000 Clumping distance cutoff}

\item{access_token}{= get_mrbase_access_token() Google OAuth2 access token. Used to authenticate level of access to data}

\item{force_server}{Force the analysis to extract results from the server rather than the MRInstruments package}

\item{force_server_if_empty}{Some of the newly added MR-Base datasets don't have pre-calculated clumped results yet. This option is soon to be deprecated but temporarily we are forcing the search for instruments when an outcome doesn't have any precalculated results. Default = TRUE.}
}
\value{
data frame
}
\description{
This function searches for GWAS significant SNPs (for a given p-value) for a specified set of outcomes.
It then performs LD based clumping to return only independent significant associations.
}
