% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{extract_instruments_gwas_catalog}
\alias{extract_instruments_gwas_catalog}
\title{Create instruments using the NHGRI-EBI GWAS Catalog}
\usage{
extract_instruments_gwas_catalog(
  trait = NULL,
  efo = NULL,
  efo_id = NULL,
  map_association_to_study = TRUE
)
}
\arguments{
\item{trait}{the trait of interest as reported in the GWAS catalog}

\item{efo}{trait of interest in the experimental factor ontology}

\item{efo_id}{identifier for the trait of interest in the experimental factor ontology}

\item{map_association_to_study}{map the genetic association results to study level information in the GWAS catalog, including GWAS catalog study identifiers and reported ancestry. Setting this argument to FALSE will speed up the function but will also result in more biased instruments and biased Mendelian randomization results. Hence it should only be set to FALSE by expert users. Default = TRUE}
}
\value{
data frame
}
\description{
This function extracts associations from the GWAS catalog and converts them into the expected exposure format for the harmonise_data function. It does not, however, check that the associations are independent. After running this function, consider using clump_data  to ensure the associations are independent. This function may take a few minutes to run, especially for traits with hundreds of genetic associations identified in multiple studies
}
