% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mysql.R
\name{extract_outcome_data}
\alias{extract_outcome_data}
\title{Extract SNP effects from GWAS summary statistics}
\usage{
extract_outcome_data(exposure_dat, outcomes, user = "mruser",
  password = "TMG_F1WnTL", dbname = "mrbase",
  host = "epi-franklin.epi.bris.ac.uk")
}
\arguments{
\item{exposure_dat}{Output from \code{read_exposure_data}}

\item{outcomes}{List of study names to search for, obtained from \code{available_outcomes}}

\item{user="mruser"}{User name}

\item{password="TMG_F1WnTL"}{Password}

\item{dbname="mrbase"}{Database}

\item{host="epi-franklin.epi.bris.ac.uk"}{Host}
}
\value{
Dataframe of summary statistics for
}
\description{
Extract SNP effects from GWAS summary statistics
}

