% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{extract_outcome_data_mysql}
\alias{extract_outcome_data_mysql}
\title{Extract SNP effects from GWAS summary statistics}
\usage{
extract_outcome_data_mysql(exposure_dat, outcomes, user, password, dbname, host,
  port)
}
\arguments{
\item{exposure_dat}{Output from \code{read_exposure_data}}

\item{outcomes}{List of study names to search for, obtained from \code{available_outcomes}}

\item{user}{User name}

\item{password}{Password}

\item{dbname}{Database}

\item{host}{Host}
}
\value{
Dataframe of summary statistics for all available outcomes
}
\description{
Any SNPs that have no beta or p value will be removed from the outcome results
}
