% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{fishers_combined_test}
\alias{fishers_combined_test}
\title{Fisher's combined test}
\usage{
fishers_combined_test(pval)
}
\arguments{
\item{pval}{Vector of outcome p-values}
}
\value{
List with the following elements:
\describe{
\item{b}{MR estimate}
\item{se}{Standard error}
\item{pval}{p-value}
}
}
\description{
Fisher's combined test
}
