% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot2.R
\name{forest_plot_basic}
\alias{forest_plot_basic}
\title{A basic forest plot}
\usage{
forest_plot_basic(dat, section = NULL, xlab = NULL, bottom = TRUE,
  trans = "identity")
}
\arguments{
\item{dat}{Output from format_mr_results()}

\item{section}{Which category in dat to plot. If NULL then prints everything}

\item{xlab}{x-axis label. Default=NULL}

\item{bottom}{Show x-axis? Default=FALSE}
}
\value{
ggplot object
}
\description{
This function is used to create a basic forest plot.
It requires the output from format_mr_results().
It is
}

