% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_1-to-many.R
\name{forest_plot_basic2}
\alias{forest_plot_basic2}
\title{A basic forest plot}
\usage{
forest_plot_basic2(dat, section = NULL, colour_group = NULL,
  colour_group_first = TRUE, xlab = NULL, bottom = TRUE,
  trans = "identity", xlim = NULL, lo = lo, up = up,
  subheading_size = subheading_size, colour_scheme = "black",
  shape_points = 15)
}
\arguments{
\item{dat}{Output from format_1_to_many()}

\item{section}{Which category in dat to plot. If NULL then prints everything}

\item{colour_group}{Which exposure to plot. If NULL then prints everything grouping by colour.}

\item{xlab}{x-axis label. Default=NULL}

\item{bottom}{Show x-axis? Default=FALSE}

\item{trans}{x-axis scale}

\item{xlim}{x-axis limits}

\item{lo}{Lower limit of x axis}

\item{up}{Upper limit of x axis}

\item{subheading_size}{text size for the subheadings. The subheadings correspond to the values of the section argument}

\item{colour_scheme}{the general colour scheme for the plot. Default is to make all text and data points black.}

\item{shape_points}{the shape of the data points to pass to geom_points(). Default is set to 15 (filled square).}
}
\value{
ggplot object
}
\description{
This function is used to create a basic forest plot.
It requires the output from format_1_to_many().
}
