% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{format_aries_mqtl}
\alias{format_aries_mqtl}
\title{Get data from methylation QTL results}
\usage{
format_aries_mqtl(aries_mqtl_subset, type = "exposure")
}
\arguments{
\item{aries_mqtl_subset}{Selected rows from \code{aries_mqtl} data loaded from \code{MRInstruments} package}

\item{type}{Are these data used as "exposure" or "outcome"? Default is "exposure"}
}
\value{
Data frame
}
\description{
See \code{format_data}
}
