% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{format_data}
\alias{format_data}
\title{Read exposure or outcome data}
\usage{
format_data(
  dat,
  type = "exposure",
  snps = NULL,
  header = TRUE,
  phenotype_col = "Phenotype",
  snp_col = "SNP",
  beta_col = "beta",
  se_col = "se",
  eaf_col = "eaf",
  effect_allele_col = "effect_allele",
  other_allele_col = "other_allele",
  pval_col = "pval",
  units_col = "units",
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  gene_col = "gene",
  id_col = "id",
  min_pval = 1e-200,
  z_col = "z",
  info_col = "info",
  chr_col = "chr",
  pos_col = "pos",
  log_pval = FALSE
)
}
\arguments{
\item{dat}{Data frame. Must have header with at least SNP column present.}

\item{type}{Is this the exposure or the outcome data that is being read in? The default is \code{"exposure"}.}

\item{snps}{SNPs to extract. If NULL then doesn't extract any and keeps all. The default is \code{NULL}.}

\item{header}{The default is \code{TRUE}.}

\item{phenotype_col}{Optional column name for the column with phenotype name corresponding the the SNP. If not present then will be created with the value \code{"Outcome"}. The default is \code{"Phenotype"}.}

\item{snp_col}{Required name of column with SNP rs IDs. The default is \code{"SNP"}.}

\item{beta_col}{Required for MR. Name of column with effect sizes. The default is \code{"beta"}.}

\item{se_col}{Required for MR. Name of column with standard errors. The default is \code{"se"}.}

\item{eaf_col}{Required for MR. Name of column with effect allele frequency. The default is \code{"eaf"}.}

\item{effect_allele_col}{Required for MR. Name of column with effect allele. Must contain only the characters "A", "C", "T" or "G". The default is \code{"effect_allele"}.}

\item{other_allele_col}{Required for MR. Name of column with non effect allele. Must contain only the characters "A", "C", "T" or "G". The default is \code{"other_allele"}.}

\item{pval_col}{Required for enrichment tests. Name of column with p-value. The default is \code{"pval"}.}

\item{units_col}{Optional column name for units. The default is \code{"units"}.}

\item{ncase_col}{Optional column name for number of cases. The default is \code{"ncase"}.}

\item{ncontrol_col}{Optional column name for number of controls. The default is \code{"ncontrol"}.}

\item{samplesize_col}{Optional column name for sample size. The default is \code{"samplesize"}.}

\item{gene_col}{Optional column name for gene name. The default is \code{"gene"}.}

\item{id_col}{The default is \code{"id"}.}

\item{min_pval}{Minimum allowed p-value. The default is \code{1e-200}.}

\item{z_col}{The default is \code{"z"}.}

\item{info_col}{The default is \code{"info_col"}.}

\item{chr_col}{The default is \code{"chr_col"}.}

\item{pos_col}{The default is \code{"pos"}.}

\item{log_pval}{The pval is -log10(P). The default is \code{FALSE}.}
}
\value{
data frame
}
\description{
Reads in exposure data. Checks and organises columns for use with MR or enrichment tests.
Infers p-values when possible from beta and se.
}
