% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{format_eqtl_mrbase}
\alias{format_eqtl_mrbase}
\title{Get data from eQTL catalog into correct format}
\usage{
format_eqtl_mrbase(eqtl_mrbase_subset, type = "exposure")
}
\arguments{
\item{eqtl_mrbase_subset}{Selected rows from \code{eqtl_mrbase} data loaded from \code{MRInstruments} package}

\item{type}{Are these data used as "exposure" or "outcome"? Default is "exposure"}
}
\value{
Data frame
}
\description{
See \code{format_data}
}

