% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{format_exposure_dat}
\alias{format_exposure_dat}
\title{Format exposure_dat into the right shape}
\usage{
format_exposure_dat(exposure_dat, exposure)
}
\arguments{
\item{exposure_dat}{Data frame}

\item{exposure}{Name of exposure trait}
}
\description{
Check that SNP is there
If SNP isn't there, error
If SNP is there, check that beta, se, eaf, a1, a2 are there
For all SNPs with those vals, say MR possible
For all without, say that MR is not possible but fisher test is
If some p values provided, fill in p-values where possible
If any SNPs not found in ensembl then remove
}

