% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{format_gtex_eqtl}
\alias{format_gtex_eqtl}
\title{Get data from eQTL catalog into correct format}
\usage{
format_gtex_eqtl(gtex_eqtl_subset, type = "exposure")
}
\arguments{
\item{gtex_eqtl_subset}{Selected rows from \code{gtex_eqtl} data loaded from \code{MRInstruments} package.}

\item{type}{Are these data used as \code{"exposure"} or \code{"outcome"}? Default is \code{"exposure"}.}
}
\value{
Data frame
}
\description{
See \code{\link[=format_data]{format_data()}}.
}
