% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{format_gwas_catalog}
\alias{format_gwas_catalog}
\title{Get data selected from GWAS catalog into correct format}
\usage{
format_gwas_catalog(gwas_catalog_subset, type = "exposure")
}
\arguments{
\item{gwas_catalog_subset}{The GWAS catalog subset.}

\item{type}{The default is \code{"exposure"}.}
}
\value{
Data frame
}
\description{
DEPRECATED. Please use \code{\link[=format_data]{format_data()}} instead.
}
\examples{
\dontrun{
require(MRInstruments)
data(gwas_catalog)
bmi <- subset(gwas_catalog, Phenotype=="Body mass index" & Year==2010 & grepl("kg", Units))
bmi <- format_data(bmi)
}
}
