% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{format_metab_qtls}
\alias{format_metab_qtls}
\title{Get data from metabolomic QTL results}
\usage{
format_metab_qtls(metab_qtls_subset, type = "exposure")
}
\arguments{
\item{metab_qtls_subset}{Selected rows from \code{metab_qtls} data loaded from \code{MRInstruments} package}

\item{type}{Are these data used as "exposure" or "outcome"? Default is "exposure"}
}
\value{
Data frame
}
\description{
See \code{format_data}
}

