% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr_results2.R
\name{generate_odds_ratios}
\alias{generate_odds_ratios}
\title{Generate odds ratios}
\usage{
generate_odds_ratios(mr_res)
}
\arguments{
\item{mr_res}{Results from mr()}
}
\value{
data frame
}
\description{
This function takes b and se from mr() and generates odds ratios and 95 percent confidence intervals
}

