% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe.R
\name{get_metrics}
\alias{get_metrics}
\title{Get metrics from summary dataset}
\usage{
get_metrics(dat)
}
\arguments{
\item{dat}{summary set}
}
\value{
data frame of metrics
}
\description{
Get metrics from summary dataset
}
