% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_mrbase_access_token}
\alias{get_mrbase_access_token}
\title{Get access token for OAuth2 access to MR Base}
\usage{
get_mrbase_access_token()
}
\value{
access token string
}
\description{
Get access token for OAuth2 access to MR Base
}
