% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rsq.r
\name{get_p_from_r2n}
\alias{get_p_from_r2n}
\title{Calculate p-value from rsq and sample size}
\usage{
get_p_from_r2n(r2, n)
}
\arguments{
\item{r2}{Rsq}

\item{n}{Sample size}
}
\value{
P-value
}
\description{
Calculate p-value from rsq and sample size
}
