% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\name{get_password}
\alias{get_password}
\title{Get password from either global options or argument}
\usage{
get_password(password)
}
\arguments{
\item{password}{String}
}
\value{
Password string, or empty string if nothing supplied
}
\description{
Get password from either global options or argument
}
