% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rsq.r
\name{get_population_allele_frequency}
\alias{get_population_allele_frequency}
\title{Estimate the allele frequency in population from case/control summary data}
\usage{
get_population_allele_frequency(af, prop, odds_ratio, prevalence)
}
\arguments{
\item{af}{Effect allele frequency (or MAF)}

\item{prop}{Proportion of samples that are cases}

\item{odds_ratio}{Odds ratio}

\item{prevalence}{Population disease prevalence}
}
\value{
Population allele frequency
}
\description{
Estimate the allele frequency in population from case/control summary data
}
