% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rsq.r
\name{get_r_from_bsen}
\alias{get_r_from_bsen}
\title{Estimate Rsq from beta, standard error and sample size}
\usage{
get_r_from_bsen(b, se, n)
}
\arguments{
\item{b}{Array of effect sizes}

\item{se}{Array of standard errors}

\item{n}{Array of (effective) sample sizes}
}
\value{
Vector of signed r values
}
\description{
Estimate Rsq from beta, standard error and sample size
}
