% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steiger.R
\name{get_r_from_lor}
\alias{get_r_from_lor}
\title{Estimate proportion of variance of liability explained by SNP in general population}
\usage{
get_r_from_lor(lor, af, ncase, ncontrol, prevalence, model = "logit")
}
\arguments{
\item{lor}{Vector of Log odds ratio}

\item{af}{Vector of allele frequencies}

\item{ncase}{Vector of Number of cases}

\item{ncontrol}{Vector of Number of controls}

\item{prevalence}{Vector of Disease prevalence in the population}

\item{model}{Is the effect size estiamted in "logit" (default) or "probit" model}
}
\description{
This uses equation 10 in Genetic Epidemiology 36 : 214–224 (2012)
}
