% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{get_se}
\alias{get_se}
\title{Get SE from effect size and pval}
\usage{
get_se(eff, pval)
}
\arguments{
\item{eff}{effect size}

\item{pval}{pvals}
}
\value{
array
}
\description{
Get SE from effect size and pval
}

