% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gwasinfo}
\alias{gwasinfo}
\title{Get list of studies with available GWAS summary statistics through API}
\usage{
gwasinfo(id = NULL, access_token = get_mrbase_access_token())
}
\arguments{
\item{id}{List of MR-Base IDs to retrieve. If NULL (default) retrieves all available datasets}

\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data}
}
\value{
Dataframe of details for all available studies
}
\description{
Get list of studies with available GWAS summary statistics through API
}
