% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{harmonise_data}
\alias{harmonise_data}
\title{Harmonise the alleles and effects between the exposure and outcome}
\usage{
harmonise_data(exposure_dat, outcome_dat, action = 2)
}
\arguments{
\item{exposure_dat}{Output from \code{read_exposure_data}.}

\item{outcome_dat}{Output from \code{extract_outcome_data}.}

\item{action}{Level of strictness in dealing with SNPs.
\itemize{
\item \code{action = 1}: Assume all alleles are coded on the forward strand, i.e. do not attempt to flip alleles
\item \code{action = 2}: Try to infer positive strand alleles, using allele frequencies for palindromes (default, conservative);
\item \code{action = 3}: Correct strand for non-palindromic SNPs, and drop all palindromic SNPs from the analysis (more conservative).
If a single value is passed then this action is applied to all outcomes.
But multiple values can be supplied as a vector, each element relating to a different outcome.
}}
}
\value{
Data frame with harmonised effects and alleles
}
\description{
In order to perform MR the effect of a SNP on an outcome and exposure must be harmonised to be
relative to the same allele.
}
\details{
Expects data in the format generated by \code{read_exposure_data} and \code{extract_outcome_data}.
This means the inputs must be dataframes with the following columns:

\code{outcome_dat}:
\itemize{
\item \code{SNP}
\item \code{beta.outcome}
\item \code{se.outcome}
\item \code{effect_allele.outcome}
\item \code{other_allele.outcome}
\item \code{eaf.outcome}
\item \code{outcome}
}

\code{exposure_dat}:
\itemize{
\item \code{SNP}
\item \code{beta.exposure}
\item \code{se.exposure}
\item \code{effect_allele.exposure}
\item \code{other_allele.exposure}
\item \code{eaf.exposure}
}

The function tries to harmonise INDELs. If they are coded as sequence strings things work more smoothly.
If they are coded as D/I in one dataset it will try to convert them to sequences if the other dataset has adequate information.
If coded as D/I in one dataset and as a variant with equal length INDEL alleles in the other, the variant is dropped.
If one or both the datasets only has one allele (i.e. the effect allele) then harmonisation is naturally going to be more ambiguous and more variants will be dropped.
}
