% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/harmonise.R
\name{harmonise_exposure_outcome}
\alias{harmonise_exposure_outcome}
\title{Harmonise the alleles and effects between the exposure and outcome}
\usage{
harmonise_exposure_outcome(exposure_dat, outcome_dat)
}
\arguments{
\item{exposure_dat}{Output from \code{read_exposure_data}}

\item{outcome_dat}{Output from \code{extract_outcome_data}}
}
\value{
Data frame with harmonised effects and alleles
}
\description{
In order to perform MR the effect of a SNP on an outcome and exposure must be harmonised to be
relative to the same allele.
}
\details{
Expects data in the format generated by \code{read_exposure_data} and \code{extract_outcome_data}.
This means the inputs must be dataframes with the following columns:
\code{outcome_dat}:
- SNP
- beta.outcome
- se.outcome
- effect_allele.outcome
- other_allele.outcome
- eaf.outcome
- outcome

\code{exposure_dat}:
- SNP
- beta.exposure
- se.exposure
- effect_allele.exposure
- other_allele.exposure
- eaf.exposure
}

