% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/harmonise.R
\name{harmonise_function}
\alias{harmonise_function}
\title{Harmonisation function}
\usage{
harmonise_function(res.tab, action)
}
\arguments{
\item{res.tab}{Data frame of exposure and outcome summary stats}
}
\value{
Data frame
}
\description{
Harmonisation function
}

