% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_formats.R
\name{harmonise_ld_dat}
\alias{harmonise_ld_dat}
\title{Harmonise LD matrix against summary data}
\usage{
harmonise_ld_dat(x, ld)
}
\arguments{
\item{x}{Exposure dataset or harmonised dataset}

\item{ld}{Output from ld_matrix}
}
\value{
List of exposure dataset and harmonised LD matrix
}
\description{
LD matrix returns with rsid_ea_oa identifiers. Make sure that they are oriented to the same effect allele as the summary dataset. Summary dataset can be exposure dataset or harmonised dartaset
}
