% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucker.R
\name{Isq}
\alias{Isq}
\title{I-squared calculation}
\usage{
Isq(y, s)
}
\arguments{
\item{y}{Vector of effects.}

\item{s}{Vector of standard errors.}
}
\value{
Isq value
}
\description{
This function calculates the \eqn{I^2} statistic.
To use it for the \eqn{I^2_{GX}} metric ensure that the effects are all the same sign (e.g. \code{abs(y)}).
}
