% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{knit_report}
\alias{knit_report}
\title{Knit report using working environment}
\usage{
knit_report(input_filename, output_filename, ...)
}
\arguments{
\item{input_filename}{Rmd file.}

\item{output_filename}{Markdown or HTML output file.  An HTML file
is specified using the .htm, .html, .HTM or .HTML file extension.
When html is specified, a similarly named markdown file is also
generated.
All output files including cache and figures will appear in the
same folder as \code{output_filename}.}

\item{...}{Arguments to be passed to \code{\link[knitr:knit]{knitr::knit}}}
}
\description{
Warning: It is quite likely that this will be called within an Rmd file
implying a recursive call to \code{knit()}. This will generate "duplicate label"
errors for unlabelled chunks. To avoid this, all code chunks
in your Rmd file should be named.
Supposedly this error can also be avoided by setting the following option:
\code{options(knitr.duplicate.label = 'allow')}.
I tried this but it didn't seem to help.
}
\keyword{internal}
