% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{ld_pruning_local}
\alias{ld_pruning_local}
\title{Perform clumping on the chosen SNPs using local data}
\usage{
ld_pruning_local(dat, refdat = NULL, clump_kb = 10000, clump_r2 = 0.1,
  clump_p1 = 1, clump_p2 = 1, plink_bin = NULL, tempdir = getwd())
}
\arguments{
\item{dat}{Output from \code{read_exposure_data}}

\item{refdat}{Path to binary plink files to use for LD referencing. If NULL (default) then looks for data provided by the Eur1000Genomes R package.}

\item{plink_bin}{Path to plink binary. If NULL (default) then looks for executable in Eur1000Genomes R package.}

\item{tempdir}{= "." Where to print results}

\item{clump_kb=10000}{Clumping window}

\item{clump_r2=0.1}{Clumping r2 cutoff}

\item{clump_p1=1}{Clumping sig level for index SNPs}

\item{clump_p2=1}{Clumping sig level for secondary SNPs}
}
\value{
Data frame of only independent SNPs
}
\description{
Perform clumping on the chosen SNPs using local data
}

