% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.r
\name{ldsc_h2_internal}
\alias{ldsc_h2_internal}
\title{Univariate LDSC}
\usage{
ldsc_h2_internal(Z, r2, N, W = NULL)
}
\arguments{
\item{Z}{summary Z-statistics for M variants}

\item{r2}{average reference LD scores for M variants}

\item{N}{GWAS sample size for each variant (could be different across variants)}

\item{W}{variant weight}
}
\value{
model fit
}
\description{
Imported here to help estimate sample overlap between studies
}
\keyword{internal}
