% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.r
\name{ldsc_rg}
\alias{ldsc_rg}
\title{Bivariate LDSC}
\usage{
ldsc_rg(id1, id2, ancestry = "infer", snpinfo = NULL, splitsize = 20000)
}
\arguments{
\item{id1}{ID 1 to analyse}

\item{id2}{ID 2 to analyse}

\item{ancestry}{ancestry of traits 1 and 2 (AFR, AMR, EAS, EUR, SAS) or 'infer' (default) in which case it will try to guess based on allele frequencies}

\item{snpinfo}{Output from ieugwasr::afl2_list("hapmap3"), or NULL for it to be done automatically}

\item{splitsize}{How many SNPs to extract at one time. Default=20000}
}
\value{
model fit
}
\description{
Imported here to help estimate sample overlap between studies
}
