% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mr.R
\name{mr}
\alias{mr}
\title{Perform all Mendelian randomization tests}
\usage{
mr(dat, bootstrap = 1000, alpha = 0.05, method_list = mr_method_list())
}
\arguments{
\item{dat}{Harmonised exposure and outcome data. Output from \code{harmonise_exposure_outcome}}

\item{bootstrap}{Number of bootstraps to estimate standard error. If NULL then don't use bootstrap}
}
\value{
List with the following elements:
        mr: Table of MR results
        extra: Table of extra results
}
\description{
Perform all Mendelian randomization tests
}

