% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_all}
\alias{mr_all}
\title{Perform Rucker, Median and Mode}
\usage{
mr_all(dat, parameters = default_parameters())
}
\arguments{
\item{dat}{Output from harmonise_data}

\item{parameters=default_parameters()}{}
}
\value{
Data frame
}
\description{
Perform Rucker, Median and Mode
}

