% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_eggers_regression}
\alias{mr_eggers_regression}
\title{Egger's regression for Mendelian randomization}
\usage{
mr_eggers_regression(b_exp, b_out, se_exp, se_out, bootstrap = NULL,
  alpha = 0.05)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}

\item{bootstrap}{Number of bootstraps to estimate standard error. If NULL then don't use bootstrap}

\item{alpha}{Quantiles to use for calculating confidence intervals using bootstraps. Default 0.05}
}
\value{
List of with the following elements:
        b: MR estimate
        se: Standard error of MR estimate
        pval: p-value of MR estimate
        b_i: Estimate of horizontal pleiotropy (intercept)
        se_i: Standard error of intercept
        pval_i: p-value of intercept
        mod: Summary of regression
        dat: Original data used for MR Egger regression
}
\description{
Egger's regression for Mendelian randomization
}

