% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{mr_forest_plot_grouped}
\alias{mr_forest_plot_grouped}
\title{Grouped forest plot}
\usage{
mr_forest_plot_grouped(
  name,
  eff_Col = "b",
  exposure_Name = "exposure",
  outcome_Name = "outcome",
  forest_Title = "",
  outfile_Name = "annot_FP.pdf",
  left_Col_Names = c("Exposure", "Outcome"),
  left_Col_Titles = NULL,
  right_Col_Names = c("p", "Outcome.n.case", "Outcome.n.control",
    "Outcome.sample.size"),
  right_Col_Titles = NULL,
  debug = FALSE,
  log_ES = FALSE,
  decrease = TRUE,
  returnRobj = TRUE,
  se_Col = "se"
)
}
\arguments{
\item{name}{(character) name of the delimited file containing all of the results on the first sheet (needs to have headers), or of the r object.}

\item{eff_Col}{(character) name of the column in the delimited file that contains the effect sizes.}

\item{exposure_Name}{(character) name of the column in the delimited file containing the \emph{types} of studies.}

\item{outcome_Name}{(character) name of the column in the delimited file containing the names of each study.}

\item{forest_Title}{(character) the title to be used for a forest plot.}

\item{outfile_Name}{(character) name to be used for output file (\emph{.pdf) or (}.wmf).}

\item{left_Col_Names}{(character vector) vector containing the names of the left-hand-side annotation columns in the delimited file.}

\item{left_Col_Titles}{(character vector) vector containing the titles for each left-hand-side annotation column.}

\item{right_Col_Names}{(character vector) vector containing the names of the right-hand-side annotation columns in the delimited file.}

\item{right_Col_Titles}{(character vector) vector containing the titles for each right-hand-side annotation column.}

\item{debug}{(logical) show warnings \code{TRUE}/\code{FALSE}?}

\item{log_ES}{(logical) perform natural log transform of effect sizes and confidence bounds \code{TRUE}/\code{FALSE}?}

\item{decrease}{(logical) sort the studies by decreasing effect sizes \code{TRUE}/\code{FALSE}?}

\item{returnRobj}{(logical) return the graph as an internal R object \code{TRUE}/\code{FALSE}?}

\item{se_Col}{(character) name of the column giving the standard error of the effect sizes.}
}
\value{
grid object giving the forest plot (or plot as pdf)
}
\description{
Grouped forest plot
}
\keyword{internal}
