% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlesnp.R
\name{mr_funnel_plot}
\alias{mr_funnel_plot}
\title{Funnel plot}
\usage{
mr_funnel_plot(singlesnp_results)
}
\arguments{
\item{singlesnp_results}{from \code{\link{mr_singlesnp}}.}
}
\value{
List of plots
}
\description{
Create funnel plot from single SNP analyses.
}
