% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mr_funnelplot}
\alias{mr_funnelplot}
\title{Funnel plot}
\usage{
mr_funnelplot(singlesnp_results)
}
\arguments{
\item{singlesnp_results}{from \code{mr_singlesnp}}
}
\value{
List of plots
}
\description{
Create funnel plot from single SNP analyses
}

