% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr-grip.R
\name{mr_grip}
\alias{mr_grip}
\title{MR-GRIP: Allele coding invariant version of MR-Egger}
\usage{
mr_grip(b_exp, b_out, se_exp, se_out, parameters)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure.}

\item{b_out}{Vector of genetic effects on outcome.}

\item{se_exp}{Standard errors of genetic effects on exposure.}

\item{se_out}{Standard errors of genetic effects on outcome.}

\item{parameters}{List of parameters.}
}
\value{
List of with the following elements:
\describe{
\item{b}{MR estimate}
\item{se}{Standard error of MR estimate}
\item{pval}{p-value of MR estimate}
\item{Q, Q_df, Q_pval}{Heterogeneity stats}
\item{b.wi}{MR estimate adjusting for weak instruments}
\item{se.wi}{Standard error adjusting for weak instruments}
\item{pval.wi}{p-value adjusting for weak instruments}
\item{mod}{Summary of regression}
\item{dat}{Original data used for MR-GRIP}
}
}
\description{
MR-GRIP: Allele coding invariant version of MR-Egger
}
