% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_ivw}
\alias{mr_ivw}
\title{Inverse variance weighted regression}
\usage{
mr_ivw(b_exp, b_out, se_exp, se_out, parameters = default_parameters())
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure.}

\item{b_out}{Vector of genetic effects on outcome.}

\item{se_exp}{Standard errors of genetic effects on exposure.}

\item{se_out}{Standard errors of genetic effects on outcome.}

\item{parameters}{List of parameters.}
}
\value{
List with the following elements:
\describe{
\item{b}{MR estimate}
\item{se}{Standard error}
\item{pval}{p-value}
\item{Q, Q_df, Q_pval}{Heterogeneity stats}
}
}
\description{
The default multiplicative random effects IVW estimate.
The standard error is corrected for under dispersion
Use the \code{\link[=mr_ivw_mre]{mr_ivw_mre()}} function for estimates that don't correct for under dispersion.
}
