% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_formats.R
\name{mr_ivw_radial}
\alias{mr_ivw_radial}
\title{Radial IVW analysis}
\usage{
mr_ivw_radial(b_exp, b_out, se_exp, se_out,
  parameters = default_parameters())
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}
}
\value{
List with the following elements:
        b: causal effect estimate
        se: standard error
        pval: p-value
}
\description{
Radial IVW analysis
}
