% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_mode}
\alias{mr_mode}
\title{MR mode estimators}
\usage{
mr_mode(dat, parameters = default_parameters())
}
\arguments{
\item{dat}{Output from harmonise_data()}

\item{parameters=default_parameters()}{<what param does>}
}
\value{
data frame
}
\description{
<full description>
}

