% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_pleiotropy_test}
\alias{mr_pleiotropy_test}
\title{Test for horizontal pleiotropy in MR analysis}
\usage{
mr_pleiotropy_test(dat)
}
\arguments{
\item{dat}{Harmonised exposure and outcome data. Output from \code{harmonise_exposure_outcome}}
}
\value{
data frame
}
\description{
Performs MR Egger and returns intercept values
}

