% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knit.R
\name{mr_report}
\alias{mr_report}
\title{Generate MR report}
\usage{
mr_report(mr_results, output_file = "mr_report.md", author = "Analyst",
  study = "Two Sample MR", ...)
}
\arguments{
\item{mr_results}{Output from \code{mr}}

\item{output_file}{Markdown or HTML output file.  An HTML file
is specified using the .htm, .html, .HTM or .HTML file extension.
When html is specified, a similarly named markdown file is also
generated.
All output files including cache and figures will appear in the
same folder as \code{output_filename}.}

\item{author}{Author name}

\item{study}{Study title}

\item{...}{Extra options to be passed to knitr}
}
\description{
Using the output from the \code{mr} function this report will generate a report containing tables and graphs summarising the results
}

