% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucker.R
\name{mr_rucker_cooksdistance}
\alias{mr_rucker_cooksdistance}
\title{MR Rucker with outliers automatically detected and removed}
\usage{
mr_rucker_cooksdistance(dat, parameters = default_parameters())
}
\arguments{
\item{dat}{Output from \code{\link{harmonise_data}}.}

\item{parameters}{List of parameters. The default is \code{default_parameters()}.}
}
\value{
List
}
\description{
Uses Cook's distance D > 4/nsnp to iteratively remove outliers.
}
