% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucker.R
\name{mr_rucker_jackknife}
\alias{mr_rucker_jackknife}
\title{Run rucker with jackknife estimates}
\usage{
mr_rucker_jackknife(dat, parameters = default_parameters())
}
\arguments{
\item{dat}{Output from harmonise_data}

\item{parameters=default_parameters()}{}
}
\value{
List
}
\description{
Run rucker with jackknife estimates
}
