% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_sign}
\alias{mr_sign}
\title{MR sign test}
\usage{
mr_sign(b_exp, b_out, se_exp = NULL, se_out = NULL, parameters = NULL)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Not required}

\item{se_out}{Not required}

\item{parameters}{Not required}
}
\value{
List with the following elements:
\describe{
\item{b}{Concordance (see description)}
\item{se}{NA}
\item{pval}{p-value}
\item{nsnp}{Number of SNPs (excludes NAs and effect estimates that are 0)}
}
}
\description{
Tests how often the SNP-exposure and SNP-outcome signs are concordant
This is to avoid the problem of averaging over all SNPs, which can suffer bias due to outliers with strong effects; and to avoid excluding SNPs which is implicit in median and mode based estimators
The effect estimate here is not to be interpreted as the effect size - it is the proportion of SNP-exposure and SNP-outcome effects that have concordant signs.
e.g. +1 means all have the same sign, -1 means all have opposite signs, and 0 means that there is an equal number of concordant and discordant signs.
Restricted to only work if there are 6 or more valid SNPs
}
