% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_simple_median}
\alias{mr_simple_median}
\title{Simple median method}
\usage{
mr_simple_median(
  b_exp,
  b_out,
  se_exp,
  se_out,
  parameters = default_parameters()
)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure.}

\item{b_out}{Vector of genetic effects on outcome.}

\item{se_exp}{Standard errors of genetic effects on exposure.}

\item{se_out}{Standard errors of genetic effects on outcome.}

\item{parameters}{The number of bootstrap replications used to calculate the SE can be set through \code{parameters=list(nboot = 1000)}. The default is \code{list(nboot=1000)}.}
}
\value{
List with the following elements:
\describe{
\item{b}{MR estimate}
\item{se}{Standard error}
\item{pval}{p-value}
\item{nsnp}{The number of SNPs}
}
}
\description{
Perform MR using summary statistics. Bootstraps used to calculate standard error.
}
