% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mode.R
\name{mr_simple_mode}
\alias{mr_simple_mode}
\title{MR simple mode estimator}
\usage{
mr_simple_mode(b_exp, b_out, se_exp, se_out, parameters = default_parameters())
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}

\item{parameters}{List containing \code{phi} - Bandwidth parameter, and \code{nboot} - number of bootstraps to calculate SE. \code{default_parameters()} sets \code{list(phi=1, nboot=1000)}.}
}
\value{
List with the following elements:
\describe{
\item{b}{MR estimate}
\item{se}{Standard error}
\item{pval}{p-value}
}
}
\description{
MR simple mode estimator
}
