% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mode.R
\name{mr_simple_mode_nome}
\alias{mr_simple_mode_nome}
\title{MR weighted mode estimator (NOME)}
\usage{
mr_simple_mode_nome(b_exp, b_out, se_exp, se_out,
  parameters = default_parameters())
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}

\item{parameters}{List containing "phi" - Bandwidth parameter, and "nboot" - number of bootstraps to calculate SE. default_parameters sets penk=1 and nboot=1000}
}
\value{
List with the following elements:
        b: MR estimate
        se: Standard error
        pval: p-value
}
\description{
MR weighted mode estimator (NOME)
}

