% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_singlesnp}
\alias{mr_singlesnp}
\title{Perform 2 sample MR on each SNP individually}
\usage{
mr_singlesnp(dat, parameters = default_parameters(),
  single_method = "mr_wald_ratio", all_method = c("mr_ivw",
  "mr_egger_regression"))
}
\arguments{
\item{dat}{Output from \code{harmonise_exposure_outcome}}

\item{parameters}{List of parameters (see default_parameters)}

\item{single_method}{Default="mr_wald_ratio" Function to use for MR analysis}

\item{all_method}{Default=c("mr_ivw", "mr_egger_regression") Function to use for MR analysis}
}
\value{
List of data frames
}
\description{
Perform 2 sample MR on each SNP individually
}

