% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_steiger}
\alias{mr_steiger}
\title{MR Steiger test of directionality}
\usage{
mr_steiger(p_exp, p_out, n_exp, n_out)
}
\arguments{
\item{p_exp}{Vector of p-values of SNP-exposure}

\item{p_out}{Vector of p-values of SNP-outcome}

\item{n_exp}{Sample sizes for p_exp}

\item{n_out}{Sample sizes for p_out}
}
\value{
List. correct_causal_direction evaluates if the exposure -> outcome assumption holds. steiger_test evaluates the confidence of the correct_causal_direction value
}
\description{
A statistical test for whether the assumption that exposure causes outcome is valid
}

