% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{mr_strategy1}
\alias{mr_strategy1}
\title{MR Strategy 1}
\usage{
mr_strategy1(dat, het_threshold = 0.05, ivw_max_snp = 1)
}
\arguments{
\item{dat}{Output from harmonise_data function}

\item{het_threshold}{The p-value threshold for Cochran's Q - if lower than this threshold then run weighted mode. Default p = 0.05}

\item{ivw_max_snp}{Maximum SNPs to allow IVW result even if heterogeneity is high. Default = 1}
}
\description{
How to choose the result for a set of different MR analysies?
Simple strategy:
Use Wald ratio if only one SNP
Use IVW if more than one SNP and heterogeneity is low
Use weighted mode if more than some minimum number of SNPs and heterogeneity is high
}
