% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_uwr}
\alias{mr_uwr}
\title{Unweighted regression}
\usage{
mr_uwr(b_exp, b_out, se_exp, se_out, parameters = default_parameters())
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}
}
\value{
List with the following elements:
        b: MR estimate
        se: Standard error
        pval: p-value
        Q, Q_df, Q_pval: Heterogeneity stats
}
\description{
The default multiplicative random effects IVW estimate. 
The standard error is corrected for under dispersion
Use the \code{mr_ivw_mre} function for estimates that don't correct for under dispersion
}
