% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_wald_ratio}
\alias{mr_wald_ratio}
\title{Perform 2 sample IV using Wald ratio.}
\usage{
mr_wald_ratio(b_exp, b_out, se_exp, se_out, parameters)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure.}

\item{b_out}{Vector of genetic effects on outcome.}

\item{se_exp}{Standard errors of genetic effects on exposure.}

\item{se_out}{Standard errors of genetic effects on outcome.}

\item{parameters}{List of parameters.}
}
\value{
List with the following elements:
\describe{
\item{b}{causal effect estimate}
\item{se}{standard error}
\item{pval}{p-value}
\item{nsnp}{1}
}
}
\description{
Perform 2 sample IV using Wald ratio.
}
