% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_weighted_median}
\alias{mr_weighted_median}
\title{Weighted median method}
\usage{
mr_weighted_median(
  b_exp,
  b_out,
  se_exp,
  se_out,
  parameters = default_parameters()
)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure.}

\item{b_out}{Vector of genetic effects on outcome.}

\item{se_exp}{Standard errors of genetic effects on exposure.}

\item{se_out}{Standard errors of genetic effects on outcome.}

\item{parameters}{The default is \code{default_parameters()}. Specify the number of bootstrap replications to calculate the SE with \code{nboot}. The default is \code{list(nboot=1000)}.}
}
\value{
List with the following elements:
\describe{
\item{b}{MR estimate}
\item{se}{Standard error}
\item{pval}{p-value}
}
}
\description{
Perform MR using summary statistics. Bootstraps used to calculate standard error.
}
