% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eve.R
\name{mr_wrapper}
\alias{mr_wrapper}
\title{Perform full set of MR analyses}
\usage{
mr_wrapper(dat)
}
\arguments{
\item{dat}{Output from \code{\link[=harmonise_data]{harmonise_data()}}.}
}
\value{
list
}
\description{
Perform full set of MR analyses
}
