% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eve.R
\name{mr_wrapper_bd}
\alias{mr_wrapper_bd}
\title{Do bi-directional MR}
\usage{
mr_wrapper_bd(ml1, ml2)
}
\arguments{
\item{ml1}{Output from read_ml}

\item{ml2}{Output from read_ml}
}
\value{
List of MR results
}
\description{
Do bi-directional MR
}
