% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R
\name{multivariable_mr}
\alias{multivariable_mr}
\title{Perform multivariable MR}
\usage{
multivariable_mr(id_exposure, id_outcome, harmonise_strictness = 2)
}
\arguments{
\item{id_exposure}{Array of ids from \code{available_outcomes}}

\item{id_outcome}{Single id from \code{available_outcomes}}
}
\value{
List of results table, exposure effects and outcome effects
}
\description{
WARNING: EXPERIMENTAL, needs more testing
1. identify exposures
2. get best instruments for each exposure
3. get effects of each instrument from each exposure and the outcome
4. create matrix of exposure effects
5. perform multivariable MR
}
