% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R
\name{mv_harmonise_data}
\alias{mv_harmonise_data}
\title{Harmonise exposure and outcome for multivariable MR}
\usage{
mv_harmonise_data(exposure_dat, outcome_dat)
}
\arguments{
\item{exposure_dat}{Output from \code{mv_extract_exposures}}

\item{outcome_dat}{Output from \code{extract_outcome_data(exposure_dat$SNP, id_output)}}
}
\value{
List of vectors and matrices required for mv analysis. exposure_beta is a matrix of beta coefficients, rows correspond to SNPs and columns correspond to exposures. exposure_pval is the same as exposure_beta, but for p-values. exposure_se is the same as exposure_beta, but for standard errors. outcome_beta is an array of effects for the outcome, corresponding to the SNPs in exposure_beta. outcome_se and outcome_pval are as in outcome_beta.
}
\description{
Harmonise exposure and outcome for multivariable MR
}

