% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R
\name{mv_lasso_feature_selection}
\alias{mv_lasso_feature_selection}
\title{Apply LASSO feature selection to mvdat object}
\usage{
mv_lasso_feature_selection(mvdat)
}
\arguments{
\item{mvdat}{Output from \code{\link[=mv_harmonise_data]{mv_harmonise_data()}}.}
}
\value{
data frame of retained features
}
\description{
Apply LASSO feature selection to mvdat object
}
