% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R
\name{mv_residual}
\alias{mv_residual}
\title{Perform basic multivariable MR}
\usage{
mv_residual(mvdat, intercept = FALSE, instrument_specific = FALSE,
  pval_threshold = 5e-08)
}
\arguments{
\item{mvdat}{Output from \code{mv_harmonise_data}}

\item{pval_threshold=5e-8}{P-value threshold to include instruments}
}
\value{
List of results
}
\description{
Performs initial multivariable MR analysis from Burgess et al 2015. For each exposure the outcome is residualised for all the other exposures, then unweighted regression is applied.
}

