% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R
\name{mv_subset}
\alias{mv_subset}
\title{Perform multivariable MR on subset of features}
\usage{
mv_subset(
  mvdat,
  features = mv_lasso_feature_selection(mvdat),
  intercept = FALSE,
  instrument_specific = FALSE,
  pval_threshold = 5e-08,
  plots = FALSE
)
}
\arguments{
\item{mvdat}{Output from \code{\link{mv_harmonise_data}}.}

\item{features}{Dataframe of features to retain, must have column with name 'exposure' that has list of exposures tor etain from mvdat. The default is \code{mvdat_lasso_feature_selection(mvdat)}.}

\item{intercept}{Should the intercept by estimated (\code{TRUE}) or force line through the origin (\code{FALSE}, the default).}

\item{instrument_specific}{Should the estimate for each exposure be obtained by using all instruments from all exposures (\code{FALSE}, default) or by using only the instruments specific to each exposure (\code{TRUE}).}

\item{pval_threshold}{P-value threshold to include instruments. The default is \code{5e-8}.}

\item{plots}{Create plots? The default is \code{FALSE}.}
}
\value{
List of results
}
\description{
The function proceeds as follows:
\enumerate{
\item Select features (by default this is done using LASSO feature selection).
\item Subset the mvdat to only retain relevant features and instruments.
\item Perform MVMR on remaining data.
}
}
