% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{outlier_adjustment}
\alias{outlier_adjustment}
\title{Outlier adjustment estimation}
\usage{
outlier_adjustment(outlierscan)
}
\arguments{
\item{outlierscan}{Output from outlier_scan}
}
\value{
data frame of adjusted effect estimates and heterogeneity stats
}
\description{
How much of the heterogeneity due to the outlier can be explained by alternative pathways?
}
