% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{outlier_graph}
\alias{outlier_graph}
\title{Plot results from outlier)scan}
\usage{
outlier_graph(outlierscan, mr_threshold_method = "fdr", mr_threshold = 0.05)
}
\arguments{
\item{outlierscan}{Output from outlier_scan function}
}
\value{
Prints plot, and returns dataframe of the connections
}
\description{
Creates a simple graph depicting the connections between outlier instruments, the original exposure and outcome traits, and the detected candidate associations
}
