% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{outlier_network}
\alias{outlier_network}
\title{Plot results from outlier_scan in a network}
\usage{
outlier_network(outlierscan)
}
\arguments{
\item{outlierscan}{Output from outlier_scan function}
}
\value{
Prints plot, and returns dataframe of the connections
}
\description{
Creates a simple network depicting the connections between outlier instruments, the original exposure and outcome traits, and the detected candidate associations
}
