% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{outlier_scan}
\alias{outlier_scan}
\title{Outlier scan}
\usage{
outlier_scan(dat, outliers = "RadialMR", use_proxies = FALSE,
  search_threshold = 5e-08, id_list = "default", include_outliers = FALSE,
  mr_method = "strategy1")
}
\arguments{
\item{dat}{Output from harmonise_data. Note - only the first id.exposure - id.outcome pair will be used}

\item{outliers}{Default is to use the RadialMR package to identify IVW outliers. Alternatively can providen an array of SNP names that are present in dat$SNP to use as outliers}

\item{use_proxies}{Whether to use proxies when looking up associations. FALSE by default for speed}

\item{search_threshold}{The p-value threshold for detecting an association between an outlier and a candidate trait. Default is 5e-8}

\item{id_list}{The list of trait IDs to search through for candidate associations. The default is the high priority traits in available_outcomes()}

\item{include_outliers}{When performing MR of candidate traits against exposures or outcomes, whether to include the original outlier SNP. Default is FALSE.}

\item{mr_method}{Method to use for candidate trait - exposure/outcome analysis. Default is strategy1. Can also provide basic MR methods e.g. mr_ivw, mr_weighted_mode etc}
}
\value{
List
dat  Cleaned dat input
radialmr  Results from RadialMR analysis
outliers  List of outliers used
id_list  List of GWAS IDs used
search  Result from search of outliers against GWAS IDs
candidate_instruments  Instruments for candidate traits
candidate_outcome  Extracted instrument SNPs from outcome
candidate_outcome_dat  Harmonised candidate - outcome dataset
candidate_outcome_mr  MR analysis of candidates against outcome
candidate_exposure   Extracted instrument SNPs from exposure
candidate_exposure_dat  Harmonised candidate - exposure dataset
candidate_exposure_mr  MR analysis of candidates against exposure
}
\description{
A simple wrapper function.
Using a summary set, find outliers in the MR analysis between the pair of trais.
Find other 'candidate traits' associated with those outliers.
Perform MR of each of those candidate traits with the original exposure and outcome
}
