% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{outlier_sig}
\alias{outlier_sig}
\title{Identify putatively significant associations in the outlier scan}
\usage{
outlier_sig(outlierscan, mr_threshold_method = "fdr", mr_threshold = 0.05)
}
\arguments{
\item{mr_threshold_method}{This is the argument to be passed to \code{p.adjust}. Default is "fdr". If no p-value adjustment is to be applied then specify "unadjusted"}

\item{mr_threshold}{Threshold to declare significance}
}
\value{
Same as outlier_scan but the candidate_exposure_mr and candidate_outcome_mr objects have an extra pval_adj and sig column each
}
\description{
Identify putatively significant associations in the outlier scan
}
