% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr_results2.R
\name{power.prune}
\alias{power.prune}
\title{Power prune}
\usage{
power.prune(dat, method.size = T)
}
\arguments{
\item{dat}{Results from harmonise_data()}

\item{method.size}{Should the duplicate summary sets be pruned on the basis of sample size? Default set to TRUE. If available, the method uses the number of cases to represent sample size. If the number of cases is missing then the total sample size is used. If method.size is set to FALSE then duplicates are dropped on the basis of expected statistical power, taking into account instrument strength and sample size, and assumes that the SNP-outcome effects are log odds ratios and that the SNP-exposure effects are in standard deviation units.}
}
\value{
data frame
}
\description{
When there are duplicate summary sets for a particular exposure-outcome combination, this function keeps the exposure-outcome summary set with the highest expected statistical power. This can be done by dropping the duplicate summary sets with the smaller outcome sample sizes. Alternatively, the pruning procedure can take into account both instrument strength and outcome sample size. The latter is useful, for example, when there is considerable variation in SNP coverage between duplicate outcome GWAS (e.g. because some studies have used targeted or fine mapping arrays). If there are a large number of SNPs available to instrument an exposure, the outcome GWAS with the better SNP coverage may provide better power than the outcome GWAS with the larger sample size. However, this latter procedure should only be implemented if the SNP-outcome effects are log odds ratios and the SNP-exposure effects are in standard deviation units.
}

