% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_exposure_data}
\alias{read_exposure_data}
\title{Read in summary statistics for exposure trait}
\usage{
read_exposure_data(filename, exposure, quote = "\\"", sep = " ")
}
\arguments{
\item{filename}{Filename}

\item{exposure}{Name of exposure trait}

\item{quote}{Character used for quotes}

\item{sep}{Character used to delimit columns}
}
\value{
Data frame of exposure summary stats
}
\description{
Read in summary statistics for exposure trait
}

