% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eve.R
\name{read_ml}
\alias{read_ml}
\title{Read in masterlist of genetic associations}
\usage{
read_ml(filename)
}
\arguments{
\item{filename}{}
}
\value{
list of two data frames
}
\description{
Read in masterlist of genetic associations
}
