% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucker.R
\name{rucker_bootstrap}
\alias{rucker_bootstrap}
\title{Run rucker with bootstrap estimates}
\usage{
rucker_bootstrap(dat, parameters = default_parameters())
}
\arguments{
\item{dat}{<what param does>}

\item{parameters=default_parameters()}{<what param does>}
}
\value{
List
}
\description{
<full description>
}

