% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mr.R
\name{run_mr}
\alias{run_mr}
\title{Perform Rucker, Median and Mode}
\usage{
run_mr(
  dat,
  parameters = default_parameters(),
  methods = c("rucker jackknife", "mode", "median"),
  plots = FALSE
)
}
\arguments{
\item{dat}{Output from harmonise_data}

\item{parameters}{default_parameters()}

\item{methods}{Named methods to use. Options are "rucker", "rucker jackknife", "rucker bootstrap", "rucker cooksdistance", "mode", "median"}
}
\value{
Data frame
}
\description{
Perform Rucker, Median and Mode
}
