% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_formats.R
\name{run_mr_presso}
\alias{run_mr_presso}
\title{Wrapper for MR-PRESSO}
\usage{
run_mr_presso(dat, NbDistribution = 1000, SignifThreshold = 0.05)
}
\arguments{
\item{dat}{Output from \code{\link{harmonise_data}}.}

\item{NbDistribution}{Number of bootstrap replications. The default is \code{1000}.}

\item{SignifThreshold}{Outlier significance threshold. The default is \code{0.05}.}
}
\value{
List of results for every exposure/outcome combination
}
\description{
See \url{https://github.com/rondolab/MR-PRESSO} for more details.
}
