% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_formats.R
\name{run_mrmix}
\alias{run_mrmix}
\title{Perform MRMix analysis on harmonised dat object}
\usage{
run_mrmix(dat)
}
\arguments{
\item{dat}{Output from harmonised_data. Ensure that eaf.exposure}
}
\value{
List of results, with one list item for every exposure/outcome pair in dat object
}
\description{
See https://github.com/gqi/MRMix for more details
}
