% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucker.R
\name{run_rucker}
\alias{run_rucker}
\title{Rucker}
\usage{
run_rucker(dat, random = "multiplicative", Qthresh = 0.05)
}
\arguments{
\item{dat}{<what param does>}

\item{random="multiplicative"}{<what param does>}

\item{Qthresh=0.05}{<what param does>}
}
\description{
<full description>
}

