% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\name{set_password}
\alias{set_password}
\title{Set password as a global option}
\usage{
set_password(password)
}
\arguments{
\item{password}{String}
}
\value{
Nothing, just prints message about whether password is correct or not, and sets global option
}
\description{
You can set the password as a global option at the beginning of a session to avoid having to write it repeatedly in analysis
}
\examples{
\dontrun{

}
}

