% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot2.R
\name{simple_cap}
\alias{simple_cap}
\title{Simple attempt at correcting string case}
\usage{
simple_cap(x)
}
\arguments{
\item{x}{Character or array of character}
}
\value{
Character or array of character
}
\description{
Simple attempt at correcting string case
}
\keyword{internal}
