% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr_results2.R
\name{size.prune}
\alias{size.prune}
\title{Size prune}
\usage{
size.prune(dat)
}
\arguments{
\item{dat}{Results from harmonise_data()}
}
\value{
data frame
}
\description{
Whens there are duplicate summary sets for a particular exposure-outcome combination, this function drops the duplicates with the smaller total sample size (for binary outcomes, the number of cases is used instead of total sample size).
}
