% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_1-to-many.R
\name{sort_1_to_many}
\alias{sort_1_to_many}
\title{Sort results for 1-to-many forest plot}
\usage{
sort_1_to_many(
  mr_res,
  b = "b",
  trait_m = "outcome",
  sort_action = 4,
  group = NULL,
  priority = NULL
)
}
\arguments{
\item{mr_res}{Data frame of results supplied by the user.}

\item{b}{Name of the column specifying the effect of the exposure on the outcome. The default is \code{"b"}.}

\item{trait_m}{The column specifying the names of the traits. Corresponds to 'many' in the 1-to-many forest plot. The default is \code{"outcome"}.}

\item{sort_action}{Choose how to sort results.
\itemize{
\item \code{sort_action = 1}: sort results by effect size within groups. Use the group order supplied by the user.
\item \code{sort_action = 2}: sort results by effect size and group. Overrides the group ordering supplied by the user.
\item \code{sort_action = 3}: group results for the same trait together (e.g. multiple results for the same trait from different MR methods).
\item \code{sort_action = 4}: sort by decreasing effect size (largest effect size at top and smallest at bottom).
\item \code{sort_action = 5}: sort by increasing effect size (smallest effect size at top and largest at bottom).
}}

\item{group}{Name of grouping variable in \code{mr_res}.}

\item{priority}{If \code{sort_action = 3}, choose which value of the \code{trait_m} variable should be given priority and go above the other \code{trait_m} values.
The trait with the largest effect size for the prioritised group will go to the top of the plot.}
}
\value{
data frame.
}
\description{
This function sorts user-supplied results for the \code{\link[=forest_plot_1_to_many]{forest_plot_1_to_many()}} function. The user supplies their results in the form of a data frame.
}
