% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr_results2.R
\name{split_exposure}
\alias{split_exposure}
\title{Split exposure column}
\usage{
split_exposure(mr_res)
}
\arguments{
\item{mr_res}{Results from mr()}
}
\value{
data frame
}
\description{
This function takes the exposure column from the results generated by mr() and splits it into separate columns for 'exposure name' and 'id'
}

