% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr_results2.R
\name{split_outcome}
\alias{split_outcome}
\title{Split outcome column}
\usage{
split_outcome(mr_res)
}
\arguments{
\item{mr_res}{Results from mr()}
}
\value{
data frame
}
\description{
This function takes the outcome column from the results generated by mr() and splits it into separate columns for 'outcome name' and 'id'
}

