% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{standardise_units}
\alias{standardise_units}
\title{Try to standardise continuous traits to be in standard deviation units}
\usage{
standardise_units(dat)
}
\arguments{
\item{dat}{Output from \code{\link{harmonise_data}}.}
}
\value{
Data frame
}
\description{
Uses \code{\link{estimate_trait_sd}}.
}
