% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steiger_filtering.R
\name{steiger_filtering}
\alias{steiger_filtering}
\title{Steiger filtering function}
\usage{
steiger_filtering(dat)
}
\arguments{
\item{dat}{Output from \code{harmonise_data}}
}
\value{
\code{harmonise_data} style data frame with additional columns rsq.exposure, rsq.outcome, steiger_dir (which is TRUE if the rsq.exposure is larger than rsq.outcome) and steiger_pval which is a test to see if rsq.exposure is significantly larger than rsq.outcome
}
\description{
This function takes an object from \code{harmonise_data} and does the following:
If there is no rsq.exposure or rsq.outcome column it will try to estimate it. This is done differently for traits that have "log odds" units. To estimate rsq for quantitative traits there must be either p-values and sample sizes for each SNP, or effect sizes and standard errors AND the units are in SD units (the column must contain "SD"). To estimate rsq for binary traits the units must be called "log odds" and there must be beta.exposure, eaf.exposure, ncase.exposure, ncontrol.exposure.
}
\details{
Once rsq is calculated for the exposure and outcome, it will then perform the steiger test for each SNP to see if the rsq of the exposure is larger than the rsq of the outcome.
}

