% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steiger.R
\name{steiger_sensitivity}
\alias{steiger_sensitivity}
\title{Evaluate the Steiger test's sensitivity to measurement error}
\usage{
steiger_sensitivity(rgx_o, rgy_o, ...)
}
\arguments{
\item{rgx_o}{Observed variance of exposure explained by SNPs}

\item{rgy_o}{Observed variance of outcome explained by SNPs}

\item{...}{Further arguments to be passed to wireframe}
}
\value{
List
- vz: Total volume of the error parameter space
- vz0: Volume of the parameter space that gives the incorrect answer
- vz1: Volume of the paramtere space that gives the correct answer
- sensitivity_ratio: Ratio of vz1/vz0. Higher means inferred direction is less susceptible to measurement error
- pl: plot of parameter sapce
}
\description{
Evaluate the Steiger test's sensitivity to measurement error
}
