% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steiger.R
\name{steiger_sensitivity_conf}
\alias{steiger_sensitivity_conf}
\title{Sensitivity analysis for unmeasured confounding on Steiger inferred causal direction}
\usage{
steiger_sensitivity_conf(
  bxy,
  bxyo,
  bgx,
  vx,
  vy,
  vg,
  vu = 1,
  simsize = 10000,
  beta_a = 1,
  beta_b = 1,
  plot = TRUE
)
}
\arguments{
\item{bxy}{MR estimate of x -> y}

\item{bxyo}{Observational estimate of x -> y, if not available can re-run with different values to test sensitivity}

\item{bgx}{SNP-exposure association}

\item{vx}{Variance of X}

\item{vy}{Variance of Y}

\item{vg}{Variance of SNP (approximately 2\emph{p}(1-p) where p is the allele frequency)}

\item{vu}{Arbitrary variance of unmeasured confounder, default = 1}

\item{simsize}{Density of search grid, default=10000}

\item{beta_a}{Weighting of confounder values, which are beta distributed. Specify 'a' parameter of beta distribution, default=1 implying flat prior}

\item{beta_b}{Weighting of confounder values, which are beta distributed. Specify 'b' parameter of beta distribution, default=1 implying flat prior}

\item{plot}{Whether to generate plot. Default=TRUE}
}
\value{
List of results
\describe{
\item{o}{data frame of possible confounding values and agreement with inferred steiger result}
\item{prop}{(weighted) fraction of confounding space that agrees with inferred steiger result}
\item{pl}{plot}
}
}
\description{
This function takes known parameters from an MR analysis and determines the range of unmeasured
confounding that would be required to agree with the Steiger test inference, and the range of
unmeasured confounding values that would be required to disagree with the Steiger test inference
}
