% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{toggle_api}
\alias{toggle_api}
\title{Toggle API address between development and release}
\usage{
toggle_api(where = "release")
}
\arguments{
\item{where}{Which API to use. Choice between "local", "release", "test". Default = "local"}
}
\description{
Toggle API address between development and release
}
