% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot2.R
\name{trim}
\alias{trim}
\title{Trim function}
\usage{
trim(x)
}
\arguments{
\item{x}{Character or array of character}
}
\value{
Character or array of character
}
\description{
Trim function to remove leading and trailing blank spaces
}
