% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{upload_file_to_api}
\alias{upload_file_to_api}
\title{Upload a file using POST request through API}
\usage{
upload_file_to_api(x, max_file_size = 16 * 1024 * 1024, header = FALSE)
}
\arguments{
\item{x}{Vector that will be written to a file to be posted}

\item{max_file_size}{Maximum file size permitted to be uploaded in bytes (16Mb default)}

\item{header}{Header in the file?}
}
\value{
basename of file
}
\description{
Upload a file using POST request through API
}

