% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variants_chrpos}
\alias{variants_chrpos}
\title{Obtain information about chr pos and surrounding region}
\usage{
variants_chrpos(chrpos, radius = 0)
}
\arguments{
\item{chrpos}{list of <chr>:<pos> in build 37, e.g. c("3:46414943", "3:122991235")}

\item{radius}{Radius around each chrpos, default = 0}
}
\value{
Data frame
}
\description{
For a list of chromosome and positions, finds all variants within a given radius
}
