% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variants_gene}
\alias{variants_gene}
\title{Obtain variants around a gene}
\usage{
variants_gene(gene, radius = 0)
}
\arguments{
\item{gene}{Vector of genes, either Ensembl or Entrez, e.g. ENSG00000123374 or 1017}

\item{radius=0}{radius around the gene region to include}
}
\value{
data frame
}
\description{
Provide a gene identified, either Ensembl or Entrez, e.g. ENSG00000123374 or 1017
}
