% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variants_rsid}
\alias{variants_rsid}
\title{Obtain information about rsid}
\usage{
variants_rsid(rsid)
}
\arguments{
\item{rsid}{Vector of rsids}
}
\value{
data frame
}
\description{
Obtain information about rsid
}
