% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{vcf_header}
\alias{vcf_header}
\title{Generate VCF header based on genome build}
\usage{
vcf_header(build = "b37")
}
\arguments{
\item{build='b37'}{Must be 'b37' or 'b38'}
}
\value{
Vector of strings to be used in vcfR package
}
\description{
Only possible to use b37 or b38 at the moment. Obtained chromosome lengths obtained frmo fai files here http://ftp.1000genomes.ebi.ac.uk/vol1/ftp/technical/reference/
}
