% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{volcano_plot}
\alias{volcano_plot}
\title{Plot volcano plot of many MR analyses}
\usage{
volcano_plot(res, what = "exposure")
}
\arguments{
\item{res}{Dataframe similar to output from mr() function, requiring id.exposure, id.outcome. Ideally only provide one MR estimate for each exposure-outcome hypothesis. Also provide a sig column of TRUE/FALSE to determine if that association is to be labelled}

\item{what}{Whether to plot many exposures against few outcomes (default: exposure) or few exposures against many outcomes (outcome). If e.g. 'exposure' and there are multiple outcomes then will facet by outcome}
}
\value{
ggplot of volcano plots
}
\description{
Plot volcano plot of many MR analyses
}
