% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{weighted_median}
\alias{weighted_median}
\title{Weighted median method}
\usage{
weighted_median(b_iv, weights)
}
\arguments{
\item{b_iv}{Wald ratios}

\item{weights}{Weights of each SNP}
}
\value{
MR estimate
}
\description{
New method from Jack
}

