% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{weighted_median_bootstrap}
\alias{weighted_median_bootstrap}
\title{Calculate standard errors for weighted median method using bootstrap}
\usage{
weighted_median_bootstrap(b_exp, b_out, se_exp, se_out, weights, nboot)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}

\item{weights}{Weights to apply to each SNP}

\item{nboot}{Number of bootstraps. Default 1000}
}
\value{
Empirical standard error
}
\description{
Based on new script for weighted median confidence interval, update 31 July 2015
}
