% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{write_vcf}
\alias{write_vcf}
\title{Write vcf file of summary GWAS data}
\usage{
write_vcf(vcf, filename)
}
\arguments{
\item{vcf}{Output from \code{make_vcf}}

\item{filename}{Must have suffix of .vcf, .vcf.gz or .bcf. If the latter two then index is also created.}
}
\description{
Writes vcf file and indexes. Note, this is only for datasets that have no genotype data. For full implementation of vcf file writing see \code{vcfR::write.vcf}. But for GWAS data it is better to use write_vcf instead of vcfR::write.vcf because the latter creates an ordinary gzip file, and we want to make a bgzip file or bcf file.
}
