% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rsq.r
\name{allele_frequency}
\alias{allele_frequency}
\title{Estimate allele frequency from SNP}
\usage{
allele_frequency(g)
}
\arguments{
\item{g}{Vector of 0/1/2}
}
\value{
Allele frequency
}
\description{
Estimate allele frequency from SNP
}
