% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{available_outcomes}
\alias{available_outcomes}
\title{Get list of studies with available GWAS summary statistics through API}
\usage{
available_outcomes(password = NULL)
}
\arguments{
\item{password}{If correct password is supplied then access to restricted studies will be supplied}
}
\value{
Dataframe of details for all available studies
}
\description{
Get list of studies with available GWAS summary statistics through API
}

