% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{available_outcomes}
\alias{available_outcomes}
\title{Get list of studies with available GWAS summary statistics through API}
\usage{
available_outcomes(access_token = ieugwasr::check_access_token())
}
\arguments{
\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data}
}
\value{
Dataframe of details for all available studies
}
\description{
Get list of studies with available GWAS summary statistics through API
}
