% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{available_outcomes}
\alias{available_outcomes}
\title{Get list of studies with available GWAS summary statistics through API}
\usage{
available_outcomes()
}
\value{
Dataframe of details for all available studies
}
\description{
Get list of studies with available GWAS summary statistics through API
}

