% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{available_outcomes}
\alias{available_outcomes}
\title{Get list of studies with available GWAS summary statistics through API}
\usage{
available_outcomes(opengwas_jwt = ieugwasr::get_opengwas_jwt())
}
\arguments{
\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to https://api.opengwas.io to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.#'}
}
\value{
Dataframe of details for all available studies
}
\description{
Get list of studies with available GWAS summary statistics through API
}
