% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{clump_data}
\alias{clump_data}
\title{Perform LD clumping on SNP data}
\usage{
clump_data(
  dat,
  clump_kb = 10000,
  clump_r2 = 0.001,
  clump_p1 = 1,
  clump_p2 = 1,
  pop = "EUR",
  bfile = NULL,
  plink_bin = NULL
)
}
\arguments{
\item{dat}{Output from \code{\link[=format_data]{format_data()}}. Must have a SNP name column (\code{SNP}), SNP chromosome column (\code{chr_name}), SNP position column (\code{chrom_start}). If \code{id.exposure} or \code{pval.exposure} not present they will be generated.}

\item{clump_kb}{Clumping window, default is \code{10000}.}

\item{clump_r2}{Clumping r2 cutoff. Note that this default value has recently changed from \code{0.01} to \code{0.001}.}

\item{clump_p1}{Clumping sig level for index SNPs, default is \code{1}.}

\item{clump_p2}{Clumping sig level for secondary SNPs, default is \code{1}.}

\item{pop}{Super-population to use as reference panel. Default = \code{"EUR"}. Options are \code{"EUR"}, \code{"SAS"}, \code{"EAS"}, \code{"AFR"}, \code{"AMR"}. \code{'legacy'} also available - which is a previously used version of the EUR panel with a slightly different set of markers}

\item{bfile}{If this is provided then will use the API. Default = \code{NULL}}

\item{plink_bin}{If \code{NULL} and \code{bfile} is not \code{NULL} then will detect packaged plink binary for specific OS. Otherwise specify path to plink binary. Default = \code{NULL}}
}
\value{
Data frame
}
\description{
Uses PLINK clumping method, where SNPs in LD within a particular window will be pruned. The SNP with the lowest p-value is retained.
}
\details{
This function interacts with the OpenGWAS API, which houses LD reference panels for the 5 super-populations in the 1000 genomes reference panel.
It includes only bi-allelic SNPs with MAF > 0.01, so it's quite possible that a variant you want to include in the clumping process will be absent.
If it is absent, it will be automatically excluded from the results.

You can check if your variants are present in the LD reference panel using \code{\link[ieugwasr:ld_reflookup]{ieugwasr::ld_reflookup()}}.

This function does put load on the OpenGWAS servers, which makes life more difficult for other users.
We have implemented a method and made available the LD reference panels to perform clumping locally, see \code{\link[ieugwasr:ld_clump]{ieugwasr::ld_clump()}} and related vignettes for details.
}
