% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{clump_data}
\alias{clump_data}
\title{Perform LD clumping on SNP data}
\usage{
clump_data(dat, clump_kb = 10000, clump_r2 = 0.001, clump_p1 = 1,
  clump_p2 = 1, access_token = get_mrbase_access_token())
}
\arguments{
\item{dat}{Output from \code{format_data}. Must have a SNP name column (SNP), SNP chromosome column (chr_name), SNP position column (chrom_start). If id.exposure or pval.exposure not present they will be generated}

\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data}

\item{clump_kb=10000}{Clumping window}

\item{clump_r2=0.001}{Clumping r2 cutoff. Note that this default value has recently changed from 0.01.}

\item{clump_p1=1}{Clumping sig level for index SNPs}

\item{clump_p2=1}{Clumping sig level for secondary SNPs}
}
\value{
Data frame
}
\description{
Uses PLINK clumping method, where SNPs in LD within a particular window will be pruned. The SNP with the lowest p-value is retained.
}
