% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{clump_data}
\alias{clump_data}
\title{Perform LD clumping on SNP data}
\usage{
clump_data(
  dat,
  clump_kb = 10000,
  clump_r2 = 0.001,
  clump_p1 = 1,
  clump_p2 = 1,
  pop = "EUR"
)
}
\arguments{
\item{dat}{Output from \code{\link{format_data}}. Must have a SNP name column (SNP), SNP chromosome column (chr_name), SNP position column (chrom_start). If id.exposure or pval.exposure not present they will be generated.}

\item{clump_kb}{Clumping window, default is \code{10000}.}

\item{clump_r2}{Clumping r2 cutoff. Note that this default value has recently changed from \code{0.01} to \code{0.001}.}

\item{clump_p1}{Clumping sig level for index SNPs, default is \code{1}.}

\item{clump_p2}{Clumping sig level for secondary SNPs, default is \code{1}.}

\item{pop}{Super-population to use as reference panel. Default = "EUR". Options are EUR, SAS, EAS, AFR, AMR. 'legacy' also available - which is a previously used verison of the EUR panel with a slightly different set of markers}
}
\value{
Data frame
}
\description{
Uses PLINK clumping method, where SNPs in LD within a particular window will be pruned. The SNP with the lowest p-value is retained.
}
