% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr_results2.R
\name{combine_all_mrresults}
\alias{combine_all_mrresults}
\title{Combine all mr results}
\usage{
combine_all_mrresults(Res, Het, Egger_intercept, SingleSNPs, ao_slc = T,
  Exp = F)
}
\arguments{
\item{Res}{Results from mr()}

\item{Het}{Results from mr_heterogeneity()}

\item{Egger_intercept}{Results from mr_pleiotropy_test()}

\item{SingleSNPs}{Results from mr_singlesnp()}

\item{ao_slc}{Logical; if set to TRUE then outcome study level characteristics are retrieved from available_outcomes(). Default is TRUE.}

\item{Exp}{Logical; if set to TRUE results are exponentiated. Useful if user wants log odds ratios expressed as odds ratios. Default is FALSE.}
}
\value{
data frame
}
\description{
This function combines results of mr(), mr_heterogeneity(), mr_pleiotropy() and mr_singlesnp() into a single data frame. It also merges the results with outcome study level characteristics in available_outcomes(). If desired it also exponentiates results (e.g. if the user wants log odds ratio converted into odds ratios with 95 percent confidence intervals)
}
