% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr_results2.R
\name{combine_all_mrresults}
\alias{combine_all_mrresults}
\title{Combine all mr results}
\usage{
combine_all_mrresults(res, het, plt, sin, ao_slc = T, exp = F,
  split_exposure = T, split_outcome = T)
}
\arguments{
\item{res}{Results from mr()}

\item{het}{Results from mr_heterogeneity()}

\item{sin}{Results from mr_singlesnp()}

\item{ao_slc}{Logical; if set to TRUE then outcome study level characteristics are retrieved from available_outcomes(). Default is TRUE.}

\item{exp}{Logical; if set to TRUE results are exponentiated. Useful if user wants log odds ratios expressed as odds ratios. Default is FALSE.}

\item{split_exposure}{Logical; if set to TRUE the exposure column is split into separate columns for the exposure name and exposure ID. Default is TRUE.}

\item{split_outcome}{Logical; if set to TRUE the outcome column is split into separate columns for the outcome name and outcome ID. Default is TRUE.}

\item{pleiotropy}{Results from mr_pleiotropy_test()}
}
\value{
data frame
}
\description{
This function combines results of mr(), mr_heterogeneity(), mr_pleiotropy_test() and mr_singlesnp() into a single data frame. It also merges the results with outcome study level characteristics in available_outcomes(). If desired it also exponentiates results (e.g. if the user wants log odds ratio converted into odds ratios with 95 percent confidence intervals). The exposure and outcome columns from the output from mr() contain both the trait names and trait ids. The combine_all_mrresults() function splits these into separate columns by default.
}
