% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{convert_outcome_to_exposure}
\alias{convert_outcome_to_exposure}
\title{Convert outcome data to exposure data}
\usage{
convert_outcome_to_exposure(outcome_dat)
}
\arguments{
\item{outcome_dat}{Output from \code{\link[=extract_outcome_data]{extract_outcome_data()}}.}
}
\value{
data frame
}
\description{
Helper function to convert results from \code{\link[=extract_outcome_data]{extract_outcome_data()}} to \code{exposure_dat} format.
}
