% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_formats.R
\name{dat_to_MRInput}
\alias{dat_to_MRInput}
\title{Convert TwoSampleMR format to MendelianRandomization format}
\usage{
dat_to_MRInput(dat, get_correlations = FALSE, pop = "EUR")
}
\arguments{
\item{dat}{Output from the \code{\link{harmonise_data}} function.}

\item{get_correlations}{Default \code{FALSE}. If \code{TRUE} then extract the LD matrix for the SNPs from the European 1000 genomes data on the MR-Base server.}

\item{pop}{If get_correlations is TRUE then use the following}
}
\value{
List of MRInput objects for each exposure/outcome combination
}
\description{
The MendelianRandomization package offers MR methods that can
be used with the same data used in the TwoSampleMR package. This
function converts from the TwoSampleMR format to the MRInput class.
}
