% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{extract_instruments}
\alias{extract_instruments}
\title{Find instruments for use in MR from the MR Base database}
\usage{
extract_instruments(
  outcomes,
  p1 = 5e-08,
  clump = TRUE,
  p2 = 5e-08,
  r2 = 0.001,
  kb = 10000,
  access_token = ieugwasr::check_access_token(),
  force_server = FALSE
)
}
\arguments{
\item{outcomes}{Array of outcome IDs (see \code{\link{available_outcomes}}).}

\item{p1}{Significance threshold. The default is \code{5e-8}.}

\item{clump}{Logical; whether to clump results. The default is \code{TRUE}.}

\item{p2}{Secondary clumping threshold. The default is \code{5e-8}.}

\item{r2}{Clumping r2 cut off. The default is \code{0.001}.}

\item{kb}{Clumping distance cutoff. The default is \code{10000}.}

\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data. The default is \code{ieugwasr::check_access_token()}.}

\item{force_server}{Force the analysis to extract results from the server rather than the MRInstruments package.}
}
\value{
data frame
}
\description{
This function searches for GWAS significant SNPs (for a given p-value) for a specified set of outcomes.
It then performs LD based clumping to return only independent significant associations.
}
