% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{extract_instruments}
\alias{extract_instruments}
\title{Find instruments for use in MR from the MR Base database}
\usage{
extract_instruments(outcomes, p1 = 5e-08, clump = TRUE, p2 = 5e-08,
  r2 = 0.1, kb = 5000, access_token = get_mrbase_access_token())
}
\arguments{
\item{outcomes}{Array of outcome IDs (see \code{available_outcomes})}

\item{p1}{= 5e-8 Significance threshold}

\item{clump}{= TRUE Whether to clump results}

\item{p2}{= 5e-8 Secondary clumping threshold}

\item{r2}{= 0.1 Clumping r2 cut off}

\item{kb}{= 5000 Clumping distance cutoff}

\item{access_token}{= get_mrbase_access_token() Google OAuth2 access token. Used to authenticate level of access to data}
}
\value{
data frame
}
\description{
This function searches for GWAS significant SNPs (for a given p-value) for a specified set of outcomes.
It then performs LD based clumping to return only independent significant associations.
}

