% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{extract_instruments}
\alias{extract_instruments}
\title{Find instruments for use in MR from the OpenGWAS database}
\usage{
extract_instruments(
  outcomes,
  p1 = 5e-08,
  clump = TRUE,
  p2 = 5e-08,
  r2 = 0.001,
  kb = 10000,
  opengwas_jwt = ieugwasr::get_opengwas_jwt(),
  force_server = FALSE
)
}
\arguments{
\item{outcomes}{Array of outcome IDs (see \code{\link[=available_outcomes]{available_outcomes()}}).}

\item{p1}{Significance threshold. The default is \code{5e-8}.}

\item{clump}{Logical; whether to clump results. The default is \code{TRUE}.}

\item{p2}{Secondary clumping threshold. The default is \code{5e-8}.}

\item{r2}{Clumping r2 cut off. The default is \code{0.001}.}

\item{kb}{Clumping distance cutoff. The default is \code{10000}.}

\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to \url{https://api.opengwas.io} to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.}

\item{force_server}{Whether to search through pre-clumped dataset or to re-extract and clump directly from the server. The default is \code{FALSE}.}
}
\value{
data frame
}
\description{
This function searches for GWAS significant SNPs (for a given p-value) for a specified set of outcomes.
It then performs LD based clumping to return only independent significant associations.
}
