% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{extract_outcome_data}
\alias{extract_outcome_data}
\title{Supply the output from \code{read_exposure_data} and all the SNPs therein will be queried against the requested outcomes in remote database using API.}
\usage{
extract_outcome_data(snps, outcomes, proxies = TRUE, rsq = 0.8,
  align_alleles = 1, palindromes = 1, maf_threshold = 0.3,
  access_token = get_mrbase_access_token())
}
\arguments{
\item{snps}{Array of SNP rs IDs}

\item{outcomes}{Array of IDs (see \code{id} column in output from \code{available_outcomes})}

\item{proxies}{Look for LD tags? Default is TRUE.}

\item{rsq}{Minimum LD rsq value (if proxies = 1). Default = 0.8.}

\item{align_alleles}{= 1 Try to align tag alleles to target alleles (if proxies = 1). 1 = yes, 0 = no}

\item{palindromes}{= 1 Allow palindromic SNPs (if proxies = 1). 1 = yes, 0 = no}

\item{maf_threshold}{= 0.3 MAF threshold to try to infer palindromic SNPs.}

\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data}
}
\value{
Dataframe of summary statistics for all available outcomes
}
\description{
Supply the output from \code{read_exposure_data} and all the SNPs therein will be queried against the requested outcomes in remote database using API.
}

