% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{extract_outcome_data}
\alias{extract_outcome_data}
\title{Supply the output from \code{\link[=read_exposure_data]{read_exposure_data()}} and all the SNPs therein will be queried against the requested outcomes in remote database using API.}
\usage{
extract_outcome_data(
  snps,
  outcomes,
  proxies = TRUE,
  rsq = 0.8,
  align_alleles = 1,
  palindromes = 1,
  maf_threshold = 0.3,
  opengwas_jwt = ieugwasr::get_opengwas_jwt(),
  splitsize = 10000,
  proxy_splitsize = 500
)
}
\arguments{
\item{snps}{Array of SNP rs IDs.}

\item{outcomes}{Array of IDs (see \code{id} column in output from \code{\link[=available_outcomes]{available_outcomes()}}).}

\item{proxies}{Look for LD tags? Default is \code{TRUE}.}

\item{rsq}{Minimum LD rsq value (if proxies = 1). Default = \code{0.8}.}

\item{align_alleles}{Try to align tag alleles to target alleles (if proxies = 1). \code{1} = yes, \code{0} = no. The default is \code{1}.}

\item{palindromes}{Allow palindromic SNPs (if proxies = 1). \code{1} = yes, \code{0} = no. The default is \code{1}.}

\item{maf_threshold}{MAF threshold to try to infer palindromic SNPs. The default is \code{0.3}.}

\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to \url{https://api.opengwas.io} to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.}

\item{splitsize}{The default is \code{10000}.}

\item{proxy_splitsize}{The default is \code{500}.}
}
\value{
Dataframe of summary statistics for all available outcomes
}
\description{
Supply the output from \code{\link[=read_exposure_data]{read_exposure_data()}} and all the SNPs therein will be queried against the requested outcomes in remote database using API.
}
