% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{extract_outcome_data_using_get}
\alias{extract_outcome_data_using_get}
\title{Supply the output from \code{read_exposure_data} and all the SNPs therein will be queried against the requested outcomes in remote database using API.
WARNING: This is unlikely to work correctly if there are a large number of SNPs}
\usage{
extract_outcome_data_using_get(snps, outcomes)
}
\arguments{
\item{outcomes}{Array of IDs (see \code{id} column in output from \code{available_outcomes})}

\item{exposure_dat}{Output from \code{read_exposure_data}}
}
\value{
Dataframe of summary statistics for all available outcomes
}
\description{
Supply the output from \code{read_exposure_data} and all the SNPs therein will be queried against the requested outcomes in remote database using API.
WARNING: This is unlikely to work correctly if there are a large number of SNPs
}
