% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mr.R
\name{fishers_combined_test}
\alias{fishers_combined_test}
\title{Fisher's combined test}
\usage{
fishers_combined_test(pval)
}
\arguments{
\item{pval}{Vector of outcome p-values}
}
\value{
List with the following elements:
        b: MR estimate
        se: Standard error
        pval: p-value
        testname: Name of the test
}
\description{
Fisher's combined test
}

