% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_1-to-many.R
\name{forest_plot_1_to_many}
\alias{forest_plot_1_to_many}
\title{1-to-many forest plot}
\usage{
forest_plot_1_to_many(mr_res, b = "b", se = "se", exponentiate = FALSE,
  trans = "identity", ao_slc = T, Lo = NULL, Up = NULL, width1 = 1,
  by = NULL, TraitM = "outcome", Alt_name = NULL)
}
\arguments{
\item{mr_res}{Data frame of results supplied by the user}

\item{b}{Name of the column specifying the effect of the exposure on the outcome. Default = "b"}

\item{se}{Name of the column specifying the standard error for b. Default = "se"}

\item{exponentiate}{Convert log odds ratios to odds ratios? Default=FALSE}

\item{trans}{Specify x-axis scale. e.g. "identity", "log2", etc. Default is "identity". If set to "identity" an additive scale is used. If set to log2 the x-axis is plotted on a multiplicative / doubling scale (preferable when plotting odds ratios and their confidence intervals).}

\item{ao_slc}{Logical; retrieve trait subcategory information using available_outcomes(). Default=FALSE}

\item{Lo}{Lower limit of X axis to plot. Must be specified by the user.}

\item{Up}{Upper limit of X axis to plot. Must be specified by the user.}

\item{by}{Name of the column indicating a grouping variable to stratify results on. Default=NULL}

\item{TraitM}{The column specifying the names of the traits. Corresponds to 'many' in the 1-to-many forest plot. Default="outcome"}

\item{Alt_name}{Name of the column specifying an alternative label for the Y axis. Typically the Y axis in a 1-to-many forest plot corresponds to a column of trait names. If, however, the user wishes to stratify the results by trait name, then an alternative column of data must be used to label the Y axis. For example, the user wishes to plot results for multiple exposures from observational and MR studies. In addition, the user wishes to group the observational and MR studies together for comparison. In this case something like a study column with vwould be used to specify Alt_name. Alt_name is only required if by=TraitM. Default = NULL.}

\item{Width1}{Width of Y axis labels. Default=1}
}
\value{
grid plot object
}
\description{
Plot results from an analysis of multiple exposures against a single outcome or a single exposure against multiple outcomes. Plots effect estimates and 95 percent confidence intervals
}
