% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_1-to-many.R
\name{forest_plot_basic2}
\alias{forest_plot_basic2}
\title{A basic forest plot}
\usage{
forest_plot_basic2(dat, section = NULL, colour_group = NULL,
  colour_group_first = TRUE, xlab = NULL, bottom = TRUE,
  trans = "identity", xlim = NULL, Lo = Lo, Up = Up)
}
\arguments{
\item{dat}{Output from format_1_to_many()}

\item{section}{Which category in dat to plot. If NULL then prints everything}

\item{colour_group}{Which exposure to plot. If NULL then prints everything grouping by colour.}

\item{xlab}{x-axis label. Default=NULL}

\item{bottom}{Show x-axis? Default=FALSE}

\item{trans}{x-axis scale}

\item{xlim}{x-axis limits}

\item{Lo}{Lower limit of x axis}

\item{Up}{Upper limit of x axis}
}
\value{
ggplot object
}
\description{
This function is used to create a basic forest plot.
It requires the output from format_1_to_many().
}
