% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{format_d}
\alias{format_d}
\title{Format the returned table from the MySQL database}
\usage{
format_d(d)
}
\arguments{
\item{d}{Data frame}
}
\value{
Data frame
}
\description{
Format the returned table from the MySQL database
}
\keyword{internal}
