% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{format_data}
\alias{format_data}
\title{Read exposure or outcome data}
\usage{
format_data(dat, type = "exposure", snps = NULL,
  phenotype_col = "Phenotype", snp_col = "SNP", beta_col = "beta",
  se_col = "se", eaf_col = "eaf", effect_allele_col = "effect_allele",
  other_allele_col = "other_allele", pval_col = "pval",
  units_col = "units", ncase_col = "ncase", ncontrol_col = "ncontrol",
  samplesize_col = "samplesize", gene_col = "gene", min_pval = 1e-200)
}
\arguments{
\item{dat}{Data frame. Must have header with at least SNP column present.}

\item{type}{Default="exposure". Is this the exposure or the outcome data that is being read in?}

\item{snps}{Default=NULL SNPs. to extract. If NULL then doesn't extract any and keeps all.}

\item{phenotype_col}{Default="Phenotype". Optional column name for the column with phenotype name corresponding the the SNP. If not present then will be created with the value "Outcome"}

\item{snp_col}{Default="SNP". Required name of column with SNP rs IDs}

\item{beta_col}{Default="beta". Required for MR. Name of column with effect sizes}

\item{se_col}{Default="se". Required for MR. Name of column with standard errors}

\item{eaf_col}{Default="eaf". Required for MR. Name of column with effect allele frequency}

\item{effect_allele_col}{Default="effect_allele". Required for MR. Name of column with effect allele. Must be "A", "C", "T" or "G"}

\item{other_allele_col}{Default="other_allele". Required for MR. Name of column with non effect allele. Must be "A", "C", "T" or "G"}

\item{pval_col}{Default="pval". Required for enrichment tests. Name of column with p-value.}

\item{units_col}{Default="units". Optional column name for units.}

\item{ncase_col}{Default="ncase". Optional column name for number of cases.}

\item{ncontrol_col}{Default="ncontrol". Optional column name for number of controls.}

\item{samplesize_col}{Default="samplesize". Optional column name for sample size.}

\item{gene_col}{Default="gene". Optional column name for gene name.}

\item{min_pval}{Default=1e-200. Minimum allowed pval}
}
\value{
data frame
}
\description{
Reads in exposure data. Checks and organises columns for use with MR or enrichment tests. Infers p-values when possible from beta and se. If it is the exposure then looks up SNPs in biomaRt to get basic info.
}

