% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_data.R
\name{format_exposure_dat}
\alias{format_exposure_dat}
\title{Format exposure_dat into the right shape}
\usage{
format_exposure_dat(exposure_dat, exposure)
}
\arguments{
\item{exposure_dat}{Data frame}

\item{exposure}{Name of exposure trait}
}
\description{
Format exposure_dat into the right shape
}

