% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rsq.r
\name{get_r_from_lor}
\alias{get_r_from_lor}
\title{Estimate proportion of variance of liability explained by SNP in general population}
\usage{
get_r_from_lor(
  lor,
  af,
  ncase,
  ncontrol,
  prevalence,
  model = "logit",
  correction = FALSE
)
}
\arguments{
\item{lor}{Vector of Log odds ratio.}

\item{af}{Vector of allele frequencies.}

\item{ncase}{Vector of Number of cases.}

\item{ncontrol}{Vector of Number of controls.}

\item{prevalence}{Vector of Disease prevalence in the population.}

\item{model}{Is the effect size estimated from the \code{"logit"} (default) or \code{"probit"} model.}

\item{correction}{Scale the estimated r by correction value. The default is \code{FALSE}.}
}
\value{
Vector of signed r values
}
\description{
This uses equation 10 in Lee et al. A Better Coefficient of Determination for Genetic Profile Analysis.
Genetic Epidemiology 36: 214–224 (2012) \url{https://doi.org/10.1002/gepi.21614}.
}
