% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rsq.r
\name{get_r_from_pn}
\alias{get_r_from_pn}
\title{Calculate variance explained from p vals and sample size}
\usage{
get_r_from_pn(p, n)
}
\arguments{
\item{p}{Array of pvals}

\item{n}{Array of sample sizes}
}
\value{
Vector of r values (all arbitrarily positive)
}
\description{
This method is an approximation, and may be numerically unstable.
Ideally you should estimate r directly from independent replication samples.
Use get_r_from_lor for binary traits.
}
