% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{harmonise_function}
\alias{harmonise_function}
\title{Harmonisation function}
\usage{
harmonise_function(res.tab, action)
}
\arguments{
\item{res.tab}{Data frame of exposure and outcome summary stats}

\item{action}{Level of strictness in dealing with SNPs. 1=Assume all reference alleles are on the positive strand, i.e. do nothing; 2=Try to infer positive strand alleles, using allele frequencies for palindromes; 3=Correct strand for non-palindromic SNPs, and drop all palindromic SNPs from the analysis}
}
\value{
Data frame
}
\description{
Harmonisation function
}
