% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe.R, R/rucker.R
\name{Isq}
\alias{Isq}
\title{Calculate I-square}
\usage{
Isq(y, s)

Isq(y, s)
}
\arguments{
\item{y}{Vector of effects.}

\item{s}{Vector of standard errors.}
}
\value{
Numeric

Isq value
}
\description{
This function calculates the I-squared_GX statistic.
}
