% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucker.R
\name{Isq}
\alias{Isq}
\title{I-square calculation}
\usage{
Isq(y, s)
}
\arguments{
\item{y}{Vector of effects.}

\item{s}{Vector of standard errors.}
}
\value{
Isq value
}
\description{
This function calculates the I-squared_GX statistic.
}
