% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{ld_matrix}
\alias{ld_matrix}
\title{Get LD matrix for list of SNPs}
\usage{
ld_matrix(snps)
}
\arguments{
\item{snps}{List of SNPs}
}
\value{
Matrix of LD r values
}
\description{
Get LD matrix for list of SNPs
}

