% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dat.R
\name{make_dat}
\alias{make_dat}
\title{Convenient function to create a harmonised dataset}
\usage{
make_dat(exposures = c(2, 301), outcomes = c(7, 1001))
}
\arguments{
\item{exposures}{default 2 and 301 (bmi and ldl)}

\item{outcomes}{default 7 and 1001 (edu and chd)}
}
\value{
Harmonised data frame
}
\description{
<full description>
}
