% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dat.R
\name{make_dat}
\alias{make_dat}
\title{Convenient function to create a harmonised dataset}
\usage{
make_dat(exposures = c(2, 301), outcomes = c(7, 1001), proxies = TRUE)
}
\arguments{
\item{exposures}{The default is \code{c(2, 301)} (BMI and LDL).}

\item{outcomes}{The default is \code{c(7, 1001)} (EDU and CHD).}

\item{proxies}{Look for proxies? Default = TRUE}
}
\value{
Harmonised data frame
}
\description{
Convenient function to create a harmonised dataset.
}
