% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{make_vcf}
\alias{make_vcf}
\title{Generate VCF format from GWAS summary dataset}
\usage{
make_vcf(
  CHROM,
  POS,
  ID,
  REF,
  ALT,
  QUAL,
  FILTER,
  B,
  SE,
  PVAL,
  AF,
  N,
  build = "b37",
  vcf_reference_name
)
}
\arguments{
\item{CHROM}{Vector of chromosome names}

\item{POS}{Vector of chromosome positions}

\item{ID}{Vector of SNP IDs}

\item{REF}{Vector of non-effect alleles - known also as reference allele}

\item{ALT}{Vector of effect alleles - known also as alternative allele}

\item{QUAL}{Vector of quality values. Can be NAs.}

\item{FILTER}{Vector of filter values. Can be NAs.}

\item{B}{Vector of effects for each SNP, relative to the ALT allele}

\item{SE}{Vector of standard errors}

\item{PVAL}{Vector}

\item{AF}{Vector of allele frequencies for the alternative allele}

\item{N}{Vector of total sample size for trait}

\item{build="b37"}{Used in CHROM and POS. This argument used to generate the header. Must be 'b37' or 'b38'.}
}
\value{
vcfR object (with empty gt slot)
}
\description{
Takes relevant vectors and produces a vcfR object. This can then be written with \code{write_vcf}.
}
