% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr}
\alias{mr}
\title{Perform all Mendelian randomization tests}
\usage{
mr(
  dat,
  parameters = default_parameters(),
  method_list = subset(mr_method_list(), use_by_default)$obj
)
}
\arguments{
\item{dat}{Harmonised exposure and outcome data. Output from \code{\link{harmonise_data}}.}

\item{parameters}{Parameters to be used for various MR methods. Default is output from \code{\link{default_parameters}}.}

\item{method_list}{List of methods to use in analysis. See \code{\link{mr_method_list}} for details.}
}
\value{
List with the following elements:
\describe{
\item{mr}{Table of MR results}
\item{extra}{Table of extra results}
}
}
\description{
Perform all Mendelian randomization tests
}
