% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mr.R
\name{mr}
\alias{mr}
\title{Perform all Mendelian randomization tests}
\usage{
mr(dat, nboot = 1000, method_list = mr_method_list())
}
\arguments{
\item{dat}{Harmonised exposure and outcome data. Output from \code{harmonise_exposure_outcome}}

\item{nboot}{Number of bootstraps to estimate standard error. Default is 1000.}

\item{method_list}{List of methods to use in analysis. See \code{mr_method_list()} for details.}
}
\value{
List with the following elements:
        mr: Table of MR results
        extra: Table of extra results
}
\description{
Perform all Mendelian randomization tests
}

