% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr}
\alias{mr}
\title{Perform all Mendelian randomization tests}
\usage{
mr(dat, parameters = default_parameters(),
  method_list = mr_method_list()$obj)
}
\arguments{
\item{dat}{Harmonised exposure and outcome data. Output from \code{harmonise_exposure_outcome}}

\item{parameters}{Parameters to be used for various MR methods. Default is output from \code{dafault_param}.}

\item{method_list}{List of methods to use in analysis. See \code{mr_method_list()} for details.}
}
\value{
List with the following elements:
        mr: Table of MR results
        extra: Table of extra results
}
\description{
Perform all Mendelian randomization tests
}

