% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_all}
\alias{mr_all}
\title{Perform Rucker, Median and Mode}
\usage{
mr_all(dat, parameters = default_parameters())
}
\arguments{
\item{dat}{<what param does>}

\item{parameters=default_parameters()}{<what param does>}
}
\value{
list
}
\description{
Perform Rucker, Median and Mode
}

