% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlesnp.R
\name{mr_density_plot}
\alias{mr_density_plot}
\title{Density plot}
\usage{
mr_density_plot(
  singlesnp_results,
  mr_results,
  exponentiate = FALSE,
  bandwidth = "nrd0"
)
}
\arguments{
\item{singlesnp_results}{from \code{mr_singlesnp}}

\item{mr_results}{Results from \code{mr}}

\item{exponentiate}{Plot on exponential scale. Default=FALSE}

\item{bandwidth}{Density bandwidth parameter}
}
\value{
List of plots
}
\description{
Density plot
}
