% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_egger_regression}
\alias{mr_egger_regression}
\title{Egger's regression for Mendelian randomization}
\usage{
mr_egger_regression(b_exp, b_out, se_exp, se_out, parameters)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}

\item{bootstrap}{Number of bootstraps to estimate standard error. If NULL then don't use bootstrap}
}
\value{
List of with the following elements:
        b: MR estimate
        se: Standard error of MR estimate
        pval: p-value of MR estimate
        b_i: Estimate of horizontal pleiotropy (intercept)
        se_i: Standard error of intercept
        pval_i: p-value of intercept
        Q, Q_df, Q_pval: Heterogeneity stats
        mod: Summary of regression
        dat: Original data used for MR Egger regression
}
\description{
Egger's regression for Mendelian randomization
}

