% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{mr_forest_plot}
\alias{mr_forest_plot}
\title{Forest plot}
\usage{
mr_forest_plot(singlesnp_results)
}
\arguments{
\item{singlesnp_results}{from \code{mr_singlesnp}}
}
\value{
List of plots
}
\description{
Forest plot
}

