% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mr_forest_plot}
\alias{mr_forest_plot}
\title{Forest plot}
\usage{
mr_forest_plot(singlesnp_results, exponentiate = FALSE)
}
\arguments{
\item{singlesnp_results}{from \code{mr_singlesnp}}

\item{exponentiate}{Plot on exponential scale. Default=FALSE}
}
\value{
List of plots
}
\description{
Forest plot
}

