% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr-grip.R
\name{mr_grip}
\alias{mr_grip}
\title{MR-GRIP: a modified MR-Egger model with the Genotype Recoding Invariance Property}
\usage{
mr_grip(b_exp, b_out, se_exp, se_out, parameters)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure.}

\item{b_out}{Vector of genetic effects on outcome.}

\item{se_exp}{Standard errors of genetic effects on exposure.}

\item{se_out}{Standard errors of genetic effects on outcome.}

\item{parameters}{List of parameters.}
}
\value{
List of with the following elements:
\describe{
\item{b}{MR estimate}
\item{se}{Standard error of MR estimate}
\item{pval}{p-value of MR estimate}
\item{Q, Q_df, Q_pval}{Heterogeneity stats}
\item{b.wi}{MR estimate adjusting for weak instruments}
\item{se.wi}{Standard error adjusting for weak instruments}
\item{pval.wi}{p-value adjusting for weak instruments}
\item{mod}{Summary of regression}
\item{dat}{Original data used for MR-GRIP}
}
}
\description{
This implements the modified MR-Egger model with the Genotype Recoding Invariance Property (MR-GRIP) due to Dudbridge and Bowden et al. (2025).
It is well known that the results of MR-Egger are sensitive to which alleles are designated as the effect alleles.
A pragmatic convention is to orient all SNPs to have positive effects on the exposure, which has some advantages in interpretation but also brings some philosophical limitations.
The MR-GRIP model is a modification to the MR-Egger model in which each term is multiplied by the genotype-phenotype associations.
This makes each term in the model invariant to allele coding.
}
