% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{mr_heterogeneity}
\alias{mr_heterogeneity}
\title{Get heterogeneity stats}
\usage{
mr_heterogeneity(
  dat,
  parameters = default_parameters(),
  method_list = subset(mr_method_list(), heterogeneity_test & use_by_default)$obj
)
}
\arguments{
\item{dat}{Harmonised exposure and outcome data. Output from \code{\link{harmonise_data}}.}

\item{parameters}{Parameters to be used for various MR methods. Default is output from \code{dafault_param}.}

\item{method_list}{List of methods to use in analysis. See \code{mr_method_list()} for details.}
}
\value{
Data frame
}
\description{
Get heterogeneity stats
}
