% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mr_leaveoneout_plot}
\alias{mr_leaveoneout_plot}
\title{Plot results from leaveoneout analysis}
\usage{
mr_leaveoneout_plot(singlesnp_results)
}
\arguments{
\item{leaveoneout_results}{Output from \code{mr_leaveoneout}}
}
\value{
List of plots
}
\description{
Plot results from leaveoneout analysis
}

