% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_median}
\alias{mr_median}
\title{MR median estimators}
\usage{
mr_median(dat, parameters = default_parameters())
}
\arguments{
\item{dat}{Output from \code{\link[=harmonise_data]{harmonise_data()}}.}

\item{parameters}{List of parameters. The default is \code{default_parameters()}.}
}
\value{
data frame
}
\description{
MR median estimators
}
