% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_median}
\alias{mr_median}
\title{MR median estimators}
\usage{
mr_median(b_exp, b_out, se_exp, se_out, parameters = default_parameters())
}
\arguments{
\item{b_exp}{<what param does>}

\item{b_out}{<what param does>}

\item{se_exp}{<what param does>}

\item{se_out}{<what param does>}

\item{parameters=default_parameters()}{<what param does>}
}
\value{
data frame
}
\description{
MR median estimators
}

