% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_meta_fixed}
\alias{mr_meta_fixed}
\title{Perform 2 sample IV using fixed effects meta analysis and delta method for standard errors}
\usage{
mr_meta_fixed(b_exp, b_out, se_exp, se_out, Cov = 0)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}
}
\value{
List with the following elements:
        b: causal effect estimate
        se: standard error
        pval: p-value
        testname: Name of test
}
\description{
Perform 2 sample IV using fixed effects meta analysis and delta method for standard errors
}

