% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mode.R
\name{mr_mode}
\alias{mr_mode}
\title{MR mode estimators}
\usage{
mr_mode(dat, parameters = default_parameters(), mode_method = "all")
}
\arguments{
\item{dat}{Output from harmonise_data()}

\item{parameters}{default_parameters()}
}
\value{
data frame
}
\description{
Perform simple, weighted, penalised modes, as well as version that use NOME assumption
}
