% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mode.R
\name{mr_mode}
\alias{mr_mode}
\title{MR mode estimators}
\usage{
mr_mode(dat, parameters = default_parameters(), mode_method = "all")
}
\arguments{
\item{dat}{Output from \code{\link[=harmonise_data]{harmonise_data()}}.}

\item{parameters}{List of parameters. The default is \code{default_parameters()}.}

\item{mode_method}{The default is \code{"all"}. The other choices are \code{'Simple mode'}, \code{'Weighted mode'}, \code{'Penalised mode'}, \code{'Simple mode (NOME)'}, \code{'Weighted mode (NOME)'}.}
}
\value{
data frame
}
\description{
Perform simple, weighted, penalised modes, as well as versions that use the NOME assumption.
}
