% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_penalised_weighted_median}
\alias{mr_penalised_weighted_median}
\title{Penalised weighted median MR}
\usage{
mr_penalised_weighted_median(
  b_exp,
  b_out,
  se_exp,
  se_out,
  parameters = default_parameters()
)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}

\item{parameters}{List containing \code{penk} - Constant term in penalisation, and \code{nboot} - number of bootstrap replications to calculate SE. \code{default_parameters()} sets \code{parameters=list(penk=20, nboot=1000)}.}
}
\value{
List with the following elements:
\describe{
\item{b}{MR estimate}
\item{se}{Standard error}
\item{pval}{p-value}
}
}
\description{
Modification to standard weighted median MR
Updated based on Burgess 2016 "Robust instrumental variable methods using multiple candidate instruments with application to Mendelian randomization"
}
