% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_penalised_weighted_median}
\alias{mr_penalised_weighted_median}
\title{Penalised weighted median MR}
\usage{
mr_penalised_weighted_median(b_exp, b_out, se_exp, se_out, penk = 20,
  nboot = 1000)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}

\item{penk}{Constant term in penalisation. Default=20}

\item{nboot}{Number of bootstraps to calculate SE. Default 1000}
}
\value{
List with the following elements:
        b: MR estimate
        se: Standard error
        pval: p-value
        testname: Name of the test
}
\description{
Modification to standard weighted median MR
}

