% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_raps}
\alias{mr_raps}
\title{Robust adjusted profile score}
\usage{
mr_raps(b_exp, b_out, se_exp, se_out, parameters = default_parameters())
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure.}

\item{b_out}{Vector of genetic effects on outcome.}

\item{se_exp}{Standard errors of genetic effects on exposure.}

\item{se_out}{Standard errors of genetic effects on outcome.}

\item{parameters}{A list of parameters. Specifically, \code{over.dispersion}, \code{loss.function}, and \code{shrinkage}:
\itemize{
\item \code{over.dispersion} is a logical concerning should the model consider overdispersion (systematic pleiotropy);
\item \code{loss.function} allows using either the squared error loss (\code{"l2"}) or robust loss functions/scores (\code{"huber"} or \code{"tukey"});
\item \code{shrinkage} is a logical specifying whether empirically partially Bayes should be used.
}

The default is \code{parameters=list(overdispersion = TRUE, loss.function = "huber", shrinkage = FALSE)}.}
}
\value{
List with the following elements:
\describe{
\item{b}{MR estimate}
\item{se}{Standard error}
\item{pval}{p-value}
\item{nsnp}{Number of SNPs}
}
}
\description{
Robust adjusted profile score
}
\details{
This function calls the \code{mr.raps} package. Please refer to the documentation of that package for more detail.
}
\references{
Qingyuan Zhao, Jingshu Wang, Jack Bowden, Dylan S. Small. Statistical inference in two-sample summary-data Mendelian randomization using robust adjusted profile score. Annals of Statistics, 48, 3, 1742--1769, \doi{10.1214/19-AOS1866}
}
