% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{mr_report}
\alias{mr_report}
\title{Generate MR report}
\usage{
mr_report(
  dat,
  output_path = ".",
  output_type = "html",
  author = "Analyst",
  study = "Two Sample MR",
  path = system.file("reports", package = "TwoSampleMR"),
  ...
)
}
\arguments{
\item{dat}{Output from \code{\link[=harmonise_data]{harmonise_data()}}}

\item{output_path}{Directory in which reports should be saved.}

\item{output_type}{Choose \code{"html"} or \code{"md"}. Default is \code{"html"}.
All output files including cache and figures will appear in the
folder specified in \code{output_path}.}

\item{author}{Author name.}

\item{study}{Study title.}

\item{path}{The filepath to the report template.}

\item{...}{Extra options to be passed to \code{\link[knitr:knit]{knitr::knit()}}.}
}
\description{
Using the output from the \code{\link[=mr]{mr()}} function this report will generate a report containing tables and graphs summarising the results.
A separate report is produced for each exposure - outcome pair that was analysed.
}
