% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knit.R
\name{mr_report}
\alias{mr_report}
\title{Generate MR report}
\usage{
mr_report(mr_results, dat, output_path = ".", output_type = "html",
  author = "Analyst", study = "Two Sample MR",
  path = system.file("reports", package = "meffil"), ...)
}
\arguments{
\item{mr_results}{Output from \code{mr}}

\item{dat}{Output from \code{harmonise_exposure_outcome}}

\item{output_path}{Directory in which reports should be saved}

\item{output_type}{Choose "html" or "md". Default is "html".
All output files including cache and figures will appear in the
folder specified in \code{output_path}.}

\item{author}{Author name}

\item{study}{Study title}

\item{...}{Extra options to be passed to knitr}
}
\description{
Using the output from the \code{mr} function this report will generate a report containing tables and graphs summarising the results.
A separate report is produced for each exposure - outcome pair that was analysed
}

