% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucker.R
\name{mr_rucker}
\alias{mr_rucker}
\title{MR Rucker framework}
\usage{
mr_rucker(dat, parameters = default_parameters())
}
\arguments{
\item{dat}{Output from \code{\link[=harmonise_data]{harmonise_data()}}.}

\item{parameters}{List of Qthresh for determining transition between models, and alpha values for calculating confidence intervals. Defaults to 0.05 for both in \code{default_parameters()}.}
}
\value{
list
}
\description{
MR Rucker framework.
}
