% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucker.R
\name{mr_rucker_bootstrap}
\alias{mr_rucker_bootstrap}
\title{Run rucker with bootstrap estimates}
\usage{
mr_rucker_bootstrap(dat, parameters = default_parameters())
}
\arguments{
\item{dat}{Output from \code{\link[=harmonise_data]{harmonise_data()}}.}

\item{parameters}{List of parameters. The default is \code{default_parameters()}.}
}
\value{
List
}
\description{
Run Rucker with bootstrap estimates.
}
