% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mr_scatter_plot}
\alias{mr_scatter_plot}
\title{Create scatter plot with lines showing the causal estimate for different MR tests}
\usage{
mr_scatter_plot(mr_results, dat)
}
\arguments{
\item{mr_results}{Output from \code{mr}}

\item{dat}{Output from \code{harmonise_exposure_outcome}}
}
\value{
List of plots
}
\description{
Requires dev version of ggplot2
}
