% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlesnp.R
\name{mr_singlesnp}
\alias{mr_singlesnp}
\title{Perform 2 sample MR on each SNP individually}
\usage{
mr_singlesnp(
  dat,
  parameters = default_parameters(),
  single_method = "mr_wald_ratio",
  all_method = c("mr_ivw", "mr_egger_regression")
)
}
\arguments{
\item{dat}{Output from \code{\link[=harmonise_data]{harmonise_data()}}.}

\item{parameters}{List of parameters. The default is \code{default_parameters()}.}

\item{single_method}{Function to use for MR analysis. The default is \code{"mr_wald_ratio"}.}

\item{all_method}{Functions to use for MR analysis. The default is \code{c("mr_ivw", "mr_egger_regression")}.}
}
\value{
List of data frames
}
\description{
Perform 2 sample MR on each SNP individually
}
