% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_singlesnp}
\alias{mr_singlesnp}
\title{Perform 2 sample MR on each SNP individually}
\usage{
mr_singlesnp(dat, parameters = default_parameters(),
  single_method = "mr_wald_ratio", all_method = c("mr_ivw",
  "mr_egger_regression"))
}
\arguments{
\item{dat}{Output from \code{harmonise_exposure_outcome}}

\item{method=mr_two_sample_ml}{Function to use for MR analysis}
}
\value{
List of data frames
}
\description{
Perform 2 sample MR on each SNP individually
}
