% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eve.R
\name{mr_wrapper}
\alias{mr_wrapper}
\title{Perform full set of MR analyses}
\usage{
mr_wrapper(dat, parameters = default_parameters())
}
\arguments{
\item{dat}{Output from \code{\link[=harmonise_data]{harmonise_data()}}.}

\item{parameters}{Parameters to pass to MR functions. Output from default_parameters() used as default.}
}
\value{
list
}
\description{
Perform full set of MR analyses
}
