% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R
\name{mv_extract_exposures}
\alias{mv_extract_exposures}
\title{Extract exposure variables for multivariable MR}
\usage{
mv_extract_exposures(id_exposure, clump_r2 = 0.001, clump_kb = 10000,
  harmonise_strictness = 2)
}
\arguments{
\item{id_exposure}{Array of IDs (e.g. c(299, 300, 302) for HDL, LDL, trigs)}

\item{clump_r2=0.01}{Once a full list of}

\item{clump_kb=10000}{<what param does>}
}
\value{
data frame in exposure_dat format
}
\description{
Requires a list of IDs from \code{available_outcomes()}. For each ID, it extracts instruments. Then, it gets the full list of all instruments and extracts those SNPs for every exposure. Finally, it keeps only the SNPs that are a) independent and b) present in all exposures, and harmonises them to be all on the same strand.
}

