% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R
\name{mv_extract_exposures}
\alias{mv_extract_exposures}
\title{Extract exposure variables for multivariable MR}
\usage{
mv_extract_exposures(
  id_exposure,
  clump_r2 = 0.001,
  clump_kb = 10000,
  harmonise_strictness = 2,
  access_token = ieugwasr::check_access_token(),
  find_proxies = TRUE,
  force_server = FALSE,
  pval_threshold = 5e-08,
  pop = "EUR"
)
}
\arguments{
\item{id_exposure}{Array of IDs (e.g. c(299, 300, 302) for HDL, LDL, trigs)}

\item{clump_r2}{The default is \code{0.01}.}

\item{clump_kb}{The default is \code{10000}.}

\item{harmonise_strictness}{See the \code{action} option of \code{\link{harmonise_data}}. The default is \code{2}.}

\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data.}

\item{find_proxies}{Look for proxies? This slows everything down but is more accurate. The default is \code{TRUE}.}

\item{force_server}{Whether to search through pre-clumped dataset or to re-extract and clump directly from the server. The default is \code{FALSE}.}

\item{pval_threshold}{Instrument detection p-value threshold. Default = 5e-8}

\item{pop}{Which 1000 genomes super population to use for clumping}
}
\value{
data frame in \code{exposure_dat} format
}
\description{
Requires a list of IDs from available_outcomes. For each ID, it extracts instruments. Then, it gets the full list of all instruments and extracts those SNPs for every exposure. Finally, it keeps only the SNPs that are a) independent and b) present in all exposures, and harmonises them to be all on the same strand.
}
