% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R
\name{mv_extract_exposures_local}
\alias{mv_extract_exposures_local}
\title{Attempt to perform MVMR using local data}
\usage{
mv_extract_exposures_local(
  filenames_exposure,
  sep = " ",
  phenotype_col = "Phenotype",
  snp_col = "SNP",
  beta_col = "beta",
  se_col = "se",
  eaf_col = "eaf",
  effect_allele_col = "effect_allele",
  other_allele_col = "other_allele",
  pval_col = "pval",
  units_col = "units",
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  gene_col = "gene",
  id_col = "id",
  min_pval = 1e-200,
  log_pval = FALSE,
  pval_threshold = 5e-08,
  clump_r2 = 0.001,
  clump_kb = 10000,
  harmonise_strictness = 2
)
}
\arguments{
\item{filenames_exposure}{Filenames for each exposure dataset. Must have header with at least SNP column present. Following arguments are used for determining how to read the filename and clumping etc.}

\item{sep}{Specify delimeter in file. The default is space, i.e. \code{sep=" "}.}

\item{phenotype_col}{Optional column name for the column with phenotype name corresponding the the SNP. If not present then will be created with the value \code{"Outcome"}. Default is \code{"Phenotype"}.}

\item{snp_col}{Required name of column with SNP rs IDs. The default is \code{"SNP"}.}

\item{beta_col}{Required for MR. Name of column with effect sizes. THe default is \code{"beta"}.}

\item{se_col}{Required for MR. Name of column with standard errors. The default is \code{"se"}.}

\item{eaf_col}{Required for MR. Name of column with effect allele frequency. The default is \code{"eaf"}.}

\item{effect_allele_col}{Required for MR. Name of column with effect allele. Must be "A", "C", "T" or "G". The default is \code{"effect_allele"}.}

\item{other_allele_col}{Required for MR. Name of column with non effect allele. Must be "A", "C", "T" or "G". The default is \code{"other_allele"}.}

\item{pval_col}{Required for enrichment tests. Name of column with p-value. The default is \code{"pval"}.}

\item{units_col}{Optional column name for units. The default is \code{"units"}.}

\item{ncase_col}{Optional column name for number of cases. The default is \code{"ncase"}.}

\item{ncontrol_col}{Optional column name for number of controls. The default is \code{"ncontrol"}.}

\item{samplesize_col}{Optional column name for sample size. The default is \code{"samplesize"}.}

\item{gene_col}{Optional column name for gene name. The default is \code{"gene"}.}

\item{id_col}{Optional column name to give the dataset an ID. Will be generated automatically if not provided for every trait / unit combination. The default is \code{"id"}.}

\item{min_pval}{Minimum allowed p-value. The default is \code{1e-200}.}

\item{log_pval}{The pval is -log10(P). The default is \code{FALSE}.}

\item{pval_threshold}{Default=5e-8 for clumping}

\item{clump_r2}{Default=0.001 for clumping}

\item{clump_kb}{Default=10000 for clumping}

\item{harmonise_strictness}{See action argument in harmonise_data. Default=2}
}
\value{
List
}
\description{
Under construction
}
