% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R
\name{mv_harmonise_data}
\alias{mv_harmonise_data}
\title{Harmonise exposure and outcome for multivariable MR}
\usage{
mv_harmonise_data(exposure_dat, outcome_dat, harmonise_strictness = 2)
}
\arguments{
\item{exposure_dat}{Output from \code{\link[=mv_extract_exposures]{mv_extract_exposures()}}.}

\item{outcome_dat}{Output from \code{extract_outcome_data(exposure_dat$SNP, id_output)}.}

\item{harmonise_strictness}{See the \code{action} option of \code{\link[=harmonise_data]{harmonise_data()}}. The default is \code{2}.}
}
\value{
List of vectors and matrices required for mv analysis.
\describe{
\item{exposure_beta}{a matrix of beta coefficients, in which rows correspond to SNPs and columns correspond to exposures.}
\item{exposure_se}{is the same as \code{exposure_beta}, but for standard errors.}
\item{exposure_pval}{the same as \code{exposure_beta}, but for p-values.}
\item{expname}{A data frame with two variables, \code{id.exposure} and \code{exposure} which are character strings.}
\item{outcome_beta}{an array of effects for the outcome, corresponding to the SNPs in \code{exposure_beta}.}
\item{outcome_se}{an array of standard errors for the outcome.}
\item{outcome_pval}{an array of p-values for the outcome.}
\item{outname}{A data frame with two variables, \code{id.outcome} and \code{outcome} which are character strings.}
}
}
\description{
Harmonise exposure and outcome for multivariable MR
}
