% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R
\name{mv_multiple}
\alias{mv_multiple}
\title{Perform IVW multivariable MR}
\usage{
mv_multiple(mvdat, intercept = FALSE, instrument_specific = FALSE,
  pval_threshold = 5e-08)
}
\arguments{
\item{mvdat}{Output from \code{mv_harmonise_data}}

\item{pval_threshold=5e-8}{P-value threshold to include instruments}
}
\value{
List of results
}
\description{
Performs modified multivariable MR analysis. For each exposure the instruments are selected then all exposures for those SNPs are regressed against the outcome together, weighting for the inverse variance of the outcome.
}
