% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable_mr.R
\name{mv_residual}
\alias{mv_residual}
\title{Perform basic multivariable MR}
\usage{
mv_residual(
  mvdat,
  intercept = FALSE,
  instrument_specific = FALSE,
  pval_threshold = 5e-08,
  plots = FALSE
)
}
\arguments{
\item{mvdat}{Output from \code{\link{mv_harmonise_data}}.}

\item{intercept}{Should the intercept by estimated (\code{TRUE}) or force line through the origin (\code{FALSE}, default).}

\item{instrument_specific}{Should the estimate for each exposure be obtained by using all instruments from all exposures (\code{FALSE}, default) or by using only the instruments specific to each exposure (\code{TRUE}).}

\item{pval_threshold}{P-value threshold to include instruments. The default is \code{5e-8}.}

\item{plots}{Create plots? The default is \code{FALSE}.}
}
\value{
List of results
}
\description{
Performs initial multivariable MR analysis from Burgess et al 2015. For each exposure the outcome is residualised for all the other exposures, then unweighted regression is applied.
}
