% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr_results2.R
\name{power.prune}
\alias{power.prune}
\title{Power prune}
\usage{
power.prune(dat, drop.duplicates = T)
}
\arguments{
\item{dat}{Results from harmonise_data()}

\item{drop.duplicates}{Should the duplicate exposure-outcome summary sets with lower power be dropped? Default set to TRUE}
}
\value{
data frame
}
\description{
Where there are duplicate summary sets for a particular exposure-outcome combination, this function identifies the exposure-outcome summary set with highest a priori statistical power, taking into account the number of cases and controls and the variance in the exposure explained by the available SNPs. The duplicate summary sets with lower power are dropped by default. The function assumes that the SNP-outcome effects are log odds ratios. The function also assumes that the SNP-exposure effects are in standard deviation units.
}
