% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr_results2.R
\name{power.prune}
\alias{power.prune}
\title{Power prune}
\usage{
power.prune(dat, drop.duplicates = T)
}
\arguments{
\item{dat}{Results from harmonise_data()}

\item{drop.duplicates}{Should duplicates with lower power be dropped? Default set to TRUE}
}
\value{
data frame
}
\description{
Where there are duplicate disease/binary outcomes identify the outcome with highest a priori statistical power, taking into account the number of cases and controls and the variance in the exposure explained by the available SNPs. The duplicate outcomes with lower power can also be dropped if requested. The function assumes that the outcome is a binary disease trait and the SNP-outcome effects are log odds ratios. The function also assumes that the SNP-exposure effects are in standard deviation units.
}
