% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_exposure_data}
\alias{read_exposure_data}
\title{Read exposure data}
\usage{
read_exposure_data(
  filename,
  clump = FALSE,
  sep = " ",
  phenotype_col = "Phenotype",
  snp_col = "SNP",
  beta_col = "beta",
  se_col = "se",
  eaf_col = "eaf",
  effect_allele_col = "effect_allele",
  other_allele_col = "other_allele",
  pval_col = "pval",
  units_col = "units",
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  gene_col = "gene",
  id_col = "id",
  min_pval = 1e-200,
  log_pval = FALSE,
  chr_col = "chr",
  pos_col = "pos"
)
}
\arguments{
\item{filename}{Filename. Must have header with at least SNP column present.}

\item{clump}{Whether to perform LD clumping with \code{\link{clump_data}} on the exposure data. The default is \code{FALSE}.}

\item{sep}{Specify delimeter in file. The default is a space, i.e. \code{" "}.}

\item{phenotype_col}{Optional column name for the column with phenotype name corresponding the the SNP. If not present then will be created with the value "Outcome". The default is \code{"Phenotype"}.}

\item{snp_col}{Required name of column with SNP rs IDs. The default is \code{"SNP"}.}

\item{beta_col}{Required for MR. Name of column with effect sizes. The default is \code{"beta"}.}

\item{se_col}{Required for MR. Name of column with standard errors. The default is \code{"se"}.}

\item{eaf_col}{Required for MR. Name of column with effect allele frequency. The default is \code{"eaf"}.}

\item{effect_allele_col}{Required for MR. Name of column with effect allele. Must be "A", "C", "T" or "G". The default is \code{"effect_allele"}.}

\item{other_allele_col}{Required for MR. Name of column with non effect allele. Must be "A", "C", "T" or "G". The default is \code{"other_allele"}.}

\item{pval_col}{Required for enrichment tests. Name of column with p-value. The default is \code{"pval"}.}

\item{units_col}{Optional column name for units. The default is \code{"units"}.}

\item{ncase_col}{Optional column name for number of cases. The default is \code{"ncase"}.}

\item{ncontrol_col}{Optional column name for number of controls. The default is \code{"ncontrol"}.}

\item{samplesize_col}{Optional column name for sample size. The default is \code{"samplesize"}.}

\item{gene_col}{Optional column name for gene name. The default is \code{"gene"}.}

\item{id_col}{Optional column name to give the dataset an ID. Will be generated automatically if not provided for every trait / unit combination. The default is \code{"id"}.}

\item{min_pval}{Minimum allowed p-value. The default is \code{1e-200}.}

\item{log_pval}{The p-value is -log10(P). The default is \code{FALSE}.}

\item{chr_col}{Optional column name for chromosome. Default is \code{"chr"}.}

\item{pos_col}{Optional column name for genetic position Default is \code{"pos"}.}
}
\value{
data frame
}
\description{
Reads in exposure data. Checks and organises columns for use with MR or enrichment tests. Infers p-values when possible from beta and se. Looks up SNPs in biomaRt to get basic info.
}
