% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_outcome_data}
\alias{read_outcome_data}
\title{Read in summary statistics for outcome trait}
\usage{
read_outcome_data(filename, outcome, quote = "\\"", sep = " ")
}
\arguments{
\item{filename}{Filename}

\item{outcome}{Name of exposure trait}

\item{quote}{Character used for quotes}

\item{sep}{Character used to delimit columns}
}
\value{
Data frame of exposure summary stats
}
\description{
Read in summary statistics for outcome trait
}

