% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{run_mr}
\alias{run_mr}
\title{Perform Rucker, Median and Mode}
\usage{
run_mr(dat, parameters = default_parameters(), methods = c("rucker",
  "rucker jackknife", "mode", "median"))
}
\arguments{
\item{dat}{Output from harmonise_data}

\item{methods}{Named methods to use. Options are "rucker", "rucker jackknife", "rucker bootstrap", "rucker cooksdistance", "mode", "median"}
}
\value{
Data frame
}
\description{
Perform Rucker, Median and Mode
}

