% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_formats.R
\name{run_mr_presso}
\alias{run_mr_presso}
\title{Wrapper for MR-PRESSO}
\usage{
run_mr_presso(dat, NbDistribution = 1000, SignifThreshold = 0.05)
}
\arguments{
\item{dat}{Output from harmonise_data}

\item{NbDistribution}{= 1000 Number of bootstraps}

\item{SignifThreshold}{= 0.05 Outlier significance threshold}
}
\value{
List of results for every exposure/outcome combination
}
\description{
See https://github.com/rondolab/MR-PRESSO
}

