% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_formats.R
\name{run_mrmix}
\alias{run_mrmix}
\title{Perform MRMix analysis on harmonised dat object}
\usage{
run_mrmix(dat)
}
\arguments{
\item{dat}{Output from \code{\link{harmonise_data}}. Ensures that no eaf.exposure values are missing.}
}
\value{
List of results, with one list item for every exposure/outcome pair in dat object
}
\description{
See \url{https://github.com/gqi/MRMix} for more details.
}
