% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_1-to-many.R
\name{sort_1_to_many}
\alias{sort_1_to_many}
\title{Sort results for 1-to-many forest plot}
\usage{
sort_1_to_many(mr_res, b = "b", sort_action = 4, group = NULL,
  priority = NULL)
}
\arguments{
\item{mr_res}{Data frame of results supplied by the user}

\item{sort_action}{Choose how to sort results. 1 =sort results by effect size within groups. Use the group order supplied by the user. 2=sort results by effect size and group. Overides the group ordering supplied by the user. 3=group results for the same trait together (e.g. multiple results for the same trait from observational and Mendelian randomization studies or from different MR methods). 4= sort by decreasing effect size (largest effect size at top and smallest at bottom). 5= sort by increasing effect size (smallest effect size at top and largest at bottom)}

\item{group}{Name of grouping variable in mr_res.}

\item{priority}{Choose which value of the grouping variable defined by the group argument should be given priority and go to the top of the plot.}
}
\value{
data frame.
}
\description{
This function sorts user-supplied results for the forest_plot_1_to_many() function. The user supplies their results in the form of a data frame.
}
