% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR-package.R
\docType{package}
\name{TwoSampleMR-package}
\alias{TwoSampleMR-package}
\alias{TwoSampleMR}
\alias{twosamplemr}
\title{TwoSampleMR: Two Sample MR functions and interface to MR Base database}
\description{
A package for performing Mendelian randomization using GWAS summary data.
It uses the \href{https://gwas.mrcieu.ac.uk/}{IEU GWAS database} to obtain data automatically,
and a wide range of methods to run the analysis. You can use the \href{https://www.mrbase.org/}{MR-Base web app}
to try out a limited range of the functionality in this package,
but for any serious work we strongly recommend using this R package.
}
\details{
\strong{Full documentation available here:} \href{https://mrcieu.github.io/TwoSampleMR/}{https://mrcieu.github.io/TwoSampleMR}
}
